Autocut

An algorithm to determine grass cut dates if none are provided. This is an abstract class and not intended for direct use. Instead, use its subclasses that implement a determine_cut() method.

The expected number of cuts is estimated from management intensity and site altitude based on data for Swiss grasslands by Huguenin et al.

References

Huguenin-Elie I, Mosimann E, Schlegel P, Lüscher A, Kessler W, Jeangros B (2017). “Grundlagen für die Düngung landwirtschaftlicher Kulturen in der Schweiz (GRUD), Kapitel 9: Düngung von Grasland.” Agrarforschung Schweiz. https://www.agrarforschungschweiz.ch/2017/06/9-duengung-von-grasland-grud-2017/.

See also

Public fields

MVS

The ModvegeSite object for which to take the cut decision.

n_cuts

Number of cuts expected for this altitude and management intensity.

Methods

Public methods

Autocut$new()

Autocut$get_expected_n_cuts()

Autocut$determine_cut()

Autocut$clone()

Method new()

Constructor

Usage

Autocut$new(MVS)

Arguments

MVS

The ModvegeSite object for which cuts shall be determined.

Get number of expected cuts

Return the number of expected cuts for a site at a given elevation and management intensity.

This uses data.frame management_parameters as a lookup table and interpolates linearly in between the specified values.

Method get_expected_n_cuts()

Usage

Autocut$get_expected_n_cuts(elevation, intensity = "high")

Arguments

elevation

The elevation of the considered site in meters above sea level.

intensity

One of ("high", "middle", "low", "extensive"). Management intensity for considered site.

Returns

Number of expected cuts per season.

Method determine_cut()

Empty method stub intended for overriding by inheriting subclasses.

Usage

Autocut$determine_cut(DOY)

Arguments

DOY

Integer day-of-the-year.

Method clone()

The objects of this class are cloneable with this method.

Usage

Autocut$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.