ModvegeParameters	R Documentation

Parameter Data Object

Description

Data structure that contains site and vegetation parameters necessary for the configuration of an rmodvege simulation run.

Parameter description

The following is a list and description of model parameters, including the vegetation parameters, which are defined through the functional group composition.

Site and model parameters

LON

geographic longitude of site in degree.

LAT

geographic latitude of site in degree.

ELV

geographic elevation of site in m.a.s.l.

WHC

water-holding capacity of site in mm.

NI

site nutritional index (dimensionless).

RUEmax

maximum radiuation use efficiency in g DM per MJ.

w_FGA

relative weight of functional group A.

w_FGB

relative weight of functional group B.

w_FGC

relative weight of functional group C.

w_FGD

relative weight of functional group D.

sigmaGV

rate of GV respirative biomass loss (dimensionless).

sigmaGR

rate of GR respirative biomass loss (dimensionless).

T0

photosynthesis activation temperature (degree C).

T1

photosynthesis plateau temperature (degree C).

T2

photosynthesis max temperature (degree C).

KGV

basic senescence rate GV (dimensionless).

KGR

basic senescence rate GR (dimensionless).

KlGV

basic abscission rate GV (dimensionless).

KlGR

basic abscission rate GR (dimensionless).

maxOMDDV

organic matter digestibility in gram per gram DV.

minOMDDR

organic matter digestibility in gram per gram DR.

Initial conditions

AgeGV

Age of green vegetative matter in degree Celsius days.

AgeGR

Age of green reproductive matter in degree Celsius days.

AgeDV

Age of dead vegetative matter in degree Celsius days.

AgeDR

Age of dead reproductive matter in degree Celsius days.

BMGV

biomass of GV (kg DM per ha).

BMGR

biomass of GR (kg DM per ha).

BMDV

biomass of DV (kg DM per ha).

BMDR

biomass of DR (kg DM per ha).

BMDR

biomass of DR (kg DM per ha).

SENGV

senescence of GV (kg DM per ha).

SENGV

senescence of GR (kg DM per ha).

ABSGV

abscission of DV (kg DM per ha).

ABSGV

abscission of DR (kg DM per ha).

ST

thermal time (degree days).

cBM

cumulative total biomass (kg per ha).

Vegetation parameters

SLA

Specific Leaf Area in m^2 per g.

pcLAM

Percentage of laminae (0, 1).

ST1

Temperature sum in degree Celsiues days after which the seasonality function SEA starts to decrease from its maximum plateau. See also SEA().

ST2

Temperature sum in degree Celsiues days after which the seasonality function SEA has decreased back to its minimum value. See also SEA().

maxSEA

Maximum value of the seasonality function SEA().

minSEA

Minimum value of the seasonality function SEA(). Usually, minSEA = 1 - (maxSEA - 1).

maxOMDGV

Maximum organic matter digestability for green vegetative matter in arbitrary units.

minOMDGV

Minimum organic matter digestability for green vegetative matter in arbitrary units.

maxOMDGR

Maximum organic matter digestability for green reproductive matter in arbitrary units.

minOMDGR

Minimum organic matter digestability for green reproductive matter in arbitrary units.

BDGV

Bulk density of green vegetative dry matter in g per m^3.

BDDV

Bulk density of dead vegetative dry matter in g per m^3.

BDGR

Bulk density of green reproductive dry matter in g per m^3.

BDDR

Bulk density of dead reproductive dry matter in g per m^3.

fg_parameter_names

Vector of strings of the variable names of all vegetation parameters governed by functional group composition.

Public fields

parameter_names

Names of all parameters and state variables.

n_parameters

Number of total parameters.

functional_group

The functional group object holding the vegetation parameters.

fg_parameter_names

Names of vegetation parameters defined by the functional group composition.

param_file

Name of the parameter file from which initial parameter values were read.

n_parameters

Number of total parameters.

Methods

Public methods

ModvegeParameters$new()

ModvegeParameters$read_parameters()

ModvegeParameters$update_functional_group()

ModvegeParameters$clone()

Method new()

Constructor

Usage

ModvegeParameters$new(param_file = NULL)

Arguments

param_file

Name of file containing the site and vegetation parameters.

Method read_parameters()

Read parameters from parameter file

Reads in parameters from the supplied param_file and stores them in internal fields.

This function carries out some basic sanity checks of the supplied param_file and reports on unidentified and missing parameter names.

Usage

ModvegeParameters$read_parameters(param_file)

Arguments

param_file

Path or name of file to read parameters from.

Returns

P List with field names as in the class variable parameter_names.

Method update_functional_group()

Update functional group parameters

Should be run whenever the functional group composition is changed in order to reflect the changes in the parameter list self$P.

Usage

ModvegeParameters$update_functional_group()

Method clone()

The objects of this class are cloneable with this method.

Usage

ModvegeParameters$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Note

Programmatically speaking, all parameters described under Parameter description are also fields of this R6Class.