ModvegeSite	R Documentation

ModvegeSite

Description

Implements the modvege grass growth model based off of Jouven et al. (2006).

This class contains model and site parameters and state variables as attributes and has methods for running modvege with weather and management input.

Use the run() method to carry out a simulation for a given year. The results are stored in the state variables in this instance and can be written to file using write_output().

Model variables

See Jouven et al. (2006) for a thorough description of all model variables.

State Variables

BM

Standing biomass in kg DM per ha.

BMG

Standing green biomass (kg DM / ha).

cBM

Cumulativeley grown biomass (kg DM / ha).

dBM

Daily grown biomass (kg DM / ha).

hvBM

Cumulative harvested biomass (kg DM / ha).

OMD

Organic matter digestibility (kg / kg).

OMDG

OMD of green matter (kg / kg).

ST

Temperature sum in degree Celsius days.

REP

Reproductive function. Gives the fraction of growth that is assigned to reproductive growth. The remainder goes into vegetative growth. Dimensionless.

PGRO

Potential growth in kg DM / ha.

GRO

Effective growth in kg DM / ha.

LAI

Leaf area index, accounting for the proportion of light intercepted by the sward. Dimensionless.

LAIGV

LAI of green vegetative biomass. Dimensionless.

AET

Actual evapotranspiration in mm.

WR

Water reserves in mm.

ENV

Function representing environmental effects on growth. Acts as a multiplicative factor. Dimensionless.

ENVfPAR

Part of ENV due to strength of incident radiation. Dimensionless.

ENVfT

Part of ENV due to temperature. Dimensionless.

ENVfW

Part of ENV due to water limitation. Dimensionless.

Initial conditions

AgeGV

Age of green vegetative matter in degree Celsius days.

AgeGR

Age of green reproductive matter in degree Celsius days.

AgeDV

Age of dead vegetative matter in degree Celsius days.

AgeDR

Age of dead reproductive matter in degree Celsius days.

BMGV

biomass of GV (kg DM per ha).

BMGR

biomass of GR (kg DM per ha).

BMDV

biomass of DV (kg DM per ha).

BMDR

biomass of DR (kg DM per ha).

BMDR

biomass of DR (kg DM per ha).

SENGV

senescence of GV (kg DM per ha).

SENGV

senescence of GR (kg DM per ha).

ABSGV

abscission of DV (kg DM per ha).

ABSGV

abscission of DR (kg DM per ha).

ST

thermal time (degree days).

cBM

cumulative total biomass (kg per ha).

Public fields

time_step

Used time step in the model in days (untested).

state_variable_namse

Vector containing the names of the model's state variables.

n_state_variables

Number of state variables.

version

Version number of the rmodvege package. Is written into output files.

site_name

Name of the site to be simulated.

run_name

Name of the simulation run. Used to distinguish between different runs at the same site.

year

Year to be simulated.

j_start_of_growing_season

Index (DOY) of the day the growing season was determined to begin.

cut_height

Height of remaining grass after cut in m.

parameters

A ModvegeParameters object.

determine_cut

Function used to decide whether a cut occurs on a given DOY. Is overloaded depending on whether management data is provided or not.

cut_DOYs

List of DOYs on which a cut occurred.

cut_during_growth_preriod

Boolean to indicate whether a cut occurred during the growth period, in which case reproductive growth is stopped.

last_DOY_for_initial_cut

autocut Start cutting after this DOY, even if yield target is not reached.

max_cut_period

autocut Maximum period to wait between subsequent cuts.

dry_precipitation_limit

autocut Maximum amount of allowed precipitation (mm) to consider a day.

dry_days_before_cut

autocut Number of days that shold be dry before a cut is made.

dry_days_after_cut

autocut Number of days that shold be dry after a cut is made.

max_cut_delay

autocut Number of days a farmer is willing to wait for dry conditions before a cut is made anyways.

cut_delays

autocut Vector to keep track of cut delay times. wait for dry conditions before a cut is made anyways.

dry_window

autocut Logical that indicates if DOY at index is considered dry enough to cut.

target_biomass

autocut Biomass amount that should to be reached by given DOY for a cut to be made.

end_of_cutting_season

autocut Determined DOY after which no more cuts are made.

BM_after_cut

autocut Amount of biomass that remains after a cut (determined through cut_height and biomass densities BDGV, BDDV, BDGR, BDDR).

weather

A WeatherData object.

management

A ManagementData object. If its is_empty field is TRUE, the autocut routine will be employed.

Active bindings

state_variable_namse

Vector containing the names of the model's state variables.

Methods

Public methods

ModvegeSite$new()

ModvegeSite$initialize_state_variables()

ModvegeSite$get_weather()

ModvegeSite$get_management()

ModvegeSite$get_start_of_growing_season()

ModvegeSite$carry_over_from_last_day()

ModvegeSite$calculate_growth()

ModvegeSite$calculate_ageing()

ModvegeSite$update_biomass()

ModvegeSite$calculate_digestibility()

ModvegeSite$apply_cuts()

ModvegeSite$determine_cut_from_input()

ModvegeSite$determine_cut_automatically()

ModvegeSite$check_dry_days()

ModvegeSite$get_target_biomass()

ModvegeSite$run()

ModvegeSite$make_header()

ModvegeSite$write_output()

ModvegeSite$set_parameters()

ModvegeSite$clone()

Method new()

Constructor

Usage

ModvegeSite$new(parameters, site_name = "-", run_name = "-")

Arguments

parameters

A ModvegeParameters object.

site_name

string Name of the simulated site.

run_name

string Name of the simulation run. Used to differentiate between different simulation conditions at the same site. Defaults to "-", which indicates no specific run name.

Method initialize_state_variables()

Initialize all relevant state variables

Firstly, an empty vector of length days_per_year is created for each variable name in the vector self$state_variable_names. Then, some additional variables relating to management and water balance are also initialized.

Usage

ModvegeSite$initialize_state_variables()

Method get_weather()

Return weather data if it exists

Usage

ModvegeSite$get_weather()

Method get_management()

Return management data if it exists

Usage

ModvegeSite$get_management()

Method get_start_of_growing_season()

Get index (or, equivalently, DOY) of the start of the growing season.

Usage

ModvegeSite$get_start_of_growing_season(
  critical_temperature = 5,
  consider_snow = FALSE,
  critical_snow = 1
)

Arguments

critical_temperature

Daily average temperature in degree Celsius required for a day being considered "warm enough" for growth.

consider_snow

Toggle whether the effect of snow cover is to be considered.

critical_snow

Minimum daily snowfall in mm for a day to be considered snowy.

Method carry_over_from_last_day()

Set calculation values to the values stored in previously used position.

Usage

ModvegeSite$carry_over_from_last_day()

Method calculate_growth()

Update all variables which change on a daily basis and calculate biomass production.

This function updates the following attributes: LAIGV AET WR ENVfPAR ENVfT ENVfW ENV PGRO GRO

Usage

ModvegeSite$calculate_growth()

Method calculate_ageing()

Calculate the ageing, senescence and abscission of different biomass compartments.

Usage

ModvegeSite$calculate_ageing()

Method update_biomass()

Calculate and update the amount of biomass in each compartment.

:TODO: These derived values can be calculated vectorially upon request.

Usage

ModvegeSite$update_biomass()

Method calculate_digestibility()

Caclulate the organic matter digestibility of each compartment.

:TODO: These derived values can be calculated vectorially upon request and thus do not need to be part of a for loop.

Usage

ModvegeSite$calculate_digestibility()

Method apply_cuts()

Simulate the effects of cuts on the grass growth.

Usage

ModvegeSite$apply_cuts()

Method determine_cut_from_input()

Read from the input whether a cut occurs on day j.

Usage

ModvegeSite$determine_cut_from_input(DOY)

Arguments

DOY

Integer day of the year for which to check.

Method determine_cut_automatically()

Decide based on simple criteria whether day of year DOY would be a good day to cut.

This follows an implementation described in Petersen, Krischan, David Kraus, Pierluigi Calanca, Mikhail A. Semenov, Klaus Butterbach-Bahl, and Ralf Kiese. “Dynamic Simulation of Management Events for Assessing Impacts of Climate Change on Pre-Alpine Grassland Productivity.” European Journal of Agronomy 128 (August 1, 2021): 126306. https://doi.org/10.1016/j.eja.2021.126306.

The decision to cut is made based on two criteria. First, it is checked whether a target biomass is reached on given DOY. The defined target depends on the DOY and is given through :func:get_target_biomass. If said biomass is present, return TRUE.

Otherwise, it is checked whether a given amount of time has passed since the last cut. Depending on whether this is the first cut of the season or not, the relevant parameters are :int:last_DOY_for_initial_cut and :int:max_cut_period. If that amount of time has passed, return TRUE, otherwise return FALSE.

Usage

ModvegeSite$determine_cut_automatically(DOY)

Arguments

DOY

Integer day of the year for which to make a cut decision.

Method check_dry_days()

Return TRUE if there are only dry days in the next n days from given DOY.

:TODO:

Usage

ModvegeSite$check_dry_days()

Method get_target_biomass()

Get target value of biomass on given DOY, which determines whether a cut is to occur.

The regression for the target biomass is based on Fig. S2 in the supplementary material of Petersen, Krischan, David Kraus, Pierluigi Calanca, Mikhail A. Semenov, Klaus Butterbach-Bahl, and Ralf Kiese. “Dynamic Simulation of Management Events for Assessing Impacts of Climate Change on Pre-Alpine Grassland Productivity.” European Journal of Agronomy 128 (August 1, 2021): 126306. https://doi.org/10.1016/j.eja.2021.126306.

A refinement to expected yield as function of altitude has been implemented according to Table 1a in Huguenen-Elie et al. "Düngung von Grasland", Agrarforschung Schweiz, 8, (6), 2017, https://www.agrarforschungschweiz.ch/2017/06/9-duengung-von-grasland-grud-2017/

Usage

ModvegeSite$get_target_biomass(DOY, intensity = "high")

Arguments

DOY

Integer day of the year to consider.

intensity

One of "high", "middle", "low" specifying management intensity..

Method run()

Carry out a ModVege simulation for one year.

Usage

ModvegeSite$run(year, weather, management)

Arguments

year

Integer specifying the year to consider. (Unused?)

weather

Weather List for given year as returned by :method:Weather$get_weather_for_year

management

Management List for given year as provided by :class:ModvegeEnvironment's :method:get_environment_for_year.

Method make_header()

Construct a string full of meta information on a ModVege run that can be inserted as a header to an output file.

Usage

ModvegeSite$make_header()

Method write_output()

Write values of ModVege results into given file.

A header with metadata is prepended to the actual data.

Usage

ModvegeSite$write_output(filename)

Arguments

filename

Path or name of filename to be created or overwritten.

Method set_parameters()

Savely update the values in self$parameters.

Special care is taken to account for potential changes to functional group weights.

:TODO: This could be implemented using just a list as input.

Usage

ModvegeSite$set_parameters(param_names, param_values)

Arguments

param_names

Vector containing the names of the parameters to be updated.

param_values

Vector containing the corresponding values to the parameters in param_names.

Method clone()

The objects of this class are cloneable with this method.

Usage

ModvegeSite$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Note

Technically, not all of the variables listed in state_variable_names are re-initialized, as some of them already hold the right values for the simulation. Particularly, this is the case for: OMDDV, OMDDR

References
\insertRef
jouven2006ModelPredictingDynamicsrmodvege

See Also

:func:get_target_biomass