Autocut based on phenology

An algorithm to determine grass cut dates if none are provided. This uses empirical data for Switzerland to determine the first and last cut dates of the season from meteorological data. The number of cuts is inferred from Huguenen et al. and these cut events are distributed equally between first and last cut dates.

References

Huguenin-Elie I, Mosimann E, Schlegel P, Lüscher A, Kessler W, Jeangros B (2017). “Grundlagen für die Düngung landwirtschaftlicher Kulturen in der Schweiz (GRUD), Kapitel 9: Düngung von Grasland.” Agrarforschung Schweiz. https://www.agrarforschungschweiz.ch/2017/06/9-duengung-von-grasland-grud-2017/.

See also

Super class

growR::Autocut -> PhenologicalAutocut

Public fields

cut_DOYs

vector containing the integer day-of-year's on which cuts occur.

Methods

Public methods

PhenologicalAutocut$new()

PhenologicalAutocut$determine_cut()

PhenologicalAutocut$clone()

Inherited methods

growR::Autocut$get_expected_n_cuts()

Method new()

Constructor

Valid cut dates are calculated upon initialization.

Usage

PhenologicalAutocut$new(MVS)

Arguments

MVS

The ModvegeSite object for which cuts shall be determined.

Method determine_cut()

Does a cut occur on DOY?

Check if DOY is in self$cut_DOYs. If so, return TRUE. Return FALSE otherwise.

Usage

PhenologicalAutocut$determine_cut(DOY)

Arguments

DOY

Integer day of the year (1-366).

Returns

Boolean

Method clone()

The objects of this class are cloneable with this method.

Usage

PhenologicalAutocut$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.