build_functional_group	R Documentation

Build the effective functional group as a weighted linear combination.

Description

Uses the weights found in :param:'P' to construct the effective functional groups and updates functional group parameters in P.

Usage

build_functional_group(P)

Arguments

P
list; name-value pairs of parameters. Has to contain w_FGX with X in (A, B, C, D).

Value

A modified list, where all functional group parameters are updated to the values of the effective functional group according to the weights w_FGX.

See Also

[FunctionalGroup()]