compare.R	R Documentation

Compare simulation results

Description

The script compare.R which ships with the rmodvege package and is automatically put into the working directory with setup_directory() can be used to compare results of rmodvege simulation runs. It is a simple script, so it can be adjusted to your personal needs.

Details

The script makes use of the packages ggplot2, patchwork and some rmodvege functions which facilitate data loading and plotting, like load_measured_data() and add_lines().