Run growR simulations

Start the loop over runs specified in the config file.

Usage

growR_run_loop(modvege_environments, output_dir = "")

Arguments

modvege_environments

A list of ModvegeEnvironment instances.

output_dir

string; name of directory to which output files are to be written. If output_dir == "" (default), no files are written.

Value

A list of the format [[run]][[year]] containing clones of the ModvegeSite instances that were run. Also write to files, if output_dir is nonempty.

Details

By default, returns an empty list but writes output to the output files as specified in the site_name and run_name fields of the supplied ModvegeEnvironment instances. Change this behaviour through the write_files and store_results arguments.

Examples

env1 = create_example_environment(site = "posieux")
env2 = create_example_environment(site = "sorens")

growR_run_loop(c(env1, env2), output_dir = "")
#> [INFO]Starting run 1 out of 2.
#> [INFO][Run 1/2]Simulating year 2013 (1/4)
#> [INFO][Run 1/2]Simulating year 2014 (2/4)
#> [INFO][Run 1/2]Simulating year 2015 (3/4)
#> [INFO][Run 1/2]Simulating year 2016 (4/4)
#> [INFO]Starting run 2 out of 2.
#> [INFO][Run 2/2]Simulating year 2013 (1/4)
#> [INFO][Run 2/2]Simulating year 2014 (2/4)
#> [INFO][Run 2/2]Simulating year 2015 (3/4)
#> [INFO][Run 2/2]Simulating year 2016 (4/4)
#> [INFO]All runs completed.
#> [[1]]
#> [[1]][[1]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.457223659360304 0.492747302067495 0.48021377988667 0.6 ...
#>     AgeDR: 502.84 504.12 504.79 508.13 513.74 517.69 519.45 519.45  ...
#>     AgeDRp: 2613.02576747072
#>     AgeDV: 502.84 504.12 504.79 508.13 513.74 517.69 519.45 519.45  ...
#>     AgeDVp: 310.006730167763
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2563.53680744978
#>     AgeGV: 102.84 104.12 104.79 108.13 113.74 117.69 119.45 119.45  ...
#>     AgeGVp: 356.062211422787
#>     BM: 747.3588 746.178324096 745.562740186702 742.500042895222 ...
#>     BMDR: 29.9148 29.876509056 29.8564917949325 29.7567711123374 2 ...
#>     BMDRp: 2.60828702924489
#>     BMDV: 297.444 296.30181504 295.70624839177 292.743271782884 28 ...
#>     BMDVp: 219.862237969144
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 470.997010495298
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     LAIGV: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     OMD: 0.681787828763536 0.681918967842231 0.681987052472049 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.866010508474576 0.865587457627119 0.865366016949152 0. ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.866010508474576 0.865587457627119 0.865366016949152 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.84 4.12 4.79 8.13 13.74 17.69 19.45 19.45 19.45 20.56  ...
#>     STp: 3324.38
#>     WR: 130 129.509787758099 129.068168429521 128.689683936513 1 ...
#>     WRp: 129.627111149557
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7235.25493639556
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 106 134 162 190 218 246 274 302
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 95
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2013
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[1]][[2]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.572832822953001 0.489080547324631 0.497218890323204 0. ...
#>     AgeDR: 502.07 505.44 508.61 512.35 515.58 517.36 524.65 532.15  ...
#>     AgeDRp: 3959.24452590726
#>     AgeDV: 502.07 505.44 508.61 512.35 515.58 517.36 524.65 532.15  ...
#>     AgeDVp: 367.259877463267
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 3070.8736642586
#>     AgeGV: 102.07 105.44 108.61 112.35 115.58 117.36 124.65 132.15  ...
#>     AgeGVp: 435.1239158176
#>     BM: 748.0749 744.959844207 742.058642727408 738.667618102169 ...
#>     BMDR: 29.9379 29.837009277 29.7424259575919 29.6311892845105 2 ...
#>     BMDRp: 0.153306987155716
#>     BMDV: 298.137 295.12283493 292.316216769816 289.036428817658 2 ...
#>     BMDVp: 202.749087683463
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 135.270954408254
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 707.307063973773
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0.458 0.5 0.454 0.184 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0.458 0.5 0.454 0.184 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     LAIGV: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     OMD: 0.681707286262378 0.682054988178933 0.682375621137396 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.866265 0.865151186440678 0.864103474576271 0.862867372 ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.866265 0.865151186440678 0.864103474576271 0.862867372 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.07 5.44 8.61 12.35 15.58 17.36 24.65 32.15 39.42 45.34 ...
#>     STp: 3719.5
#>     WR: 129.946718652693 130 130 130 130 129.365338865 128.74430 ...
#>     WRp: 127.011229396472
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 11463.8028078937
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 91 119 147 175 203 231 259 287
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 64
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2014
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[1]][[3]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.00985649314933525
#>     ABSDV: 1.83842313064963
#>     AET: 0.19920064755 0.19901996769842 0.592248295151082 0.81323 ...
#>     AgeDR: 500 500 504.91 509.88 509.88 509.88 510.27 515.11 522.86 ...
#>     AgeDRp: 2950.65995373358
#>     AgeDV: 500 500 504.91 509.88 509.88 509.88 510.27 515.11 522.86 ...
#>     AgeDVp: 320.820532488754
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2951.13482002105
#>     AgeGV: 100 100 104.91 109.88 109.88 109.88 110.27 115.11 122.86 ...
#>     AgeGVp: 338.246964126842
#>     BM: 750 750 745.4337 740.878219371 740.878219371 740.8782193 ...
#>     BMDR: 30 30 29.8527 29.704332081 29.704332081 29.704332081 29. ...
#>     BMDRp: 1.55342681628609
#>     BMDV: 300 300 295.581 291.17388729 291.17388729 291.17388729 2 ...
#>     BMDVp: 217.307698658349
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 509.617113025523
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0 0 0.55 1 0 0 0.084 0.224 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0.55 1 0 0 0.084 0.224 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     LAIGV: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     OMD: 0.681491525423729 0.681491525423729 0.682005119290321 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.866949152542373 0.866949152542373 0.865326355932203 0. ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.866949152542373 0.866949152542373 0.865326355932203 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 0 0 4.91 9.88 9.88 9.88 10.27 15.11 22.86 35.36 40.23 42 ...
#>     STp: 3711.49
#>     WR: 129.800799840891 129.603626755415 130 129.709778732205 1 ...
#>     WRp: 117.9543167997
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7396.84166354848
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 90 118 146 174 202 230 258 286 314
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 75
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2015
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[1]][[4]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.272542889414208 0.224447085399936 0.288574824085632 0. ...
#>     AgeDR: 503.99 507.78 511.31 516.04 520.55 523.86 527.42 531.52  ...
#>     AgeDRp: 2801.12465101814
#>     AgeDV: 503.99 507.78 511.31 516.04 520.55 523.86 527.42 531.52  ...
#>     AgeDVp: 339.499460100257
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2754.74386558187
#>     AgeGV: 103.99 107.78 111.31 116.04 120.55 123.86 127.42 131.52  ...
#>     AgeGVp: 383.05172521267
#>     BM: 746.2893 742.805883333 739.597524427364 735.343037848355 ...
#>     BMDR: 29.8803 29.767053663 29.6619759635696 29.5216748172619 2 ...
#>     BMDRp: 1.94108326472162
#>     BMDV: 296.409 293.03882967 289.935548463795 285.821363031093 2 ...
#>     BMDVp: 209.232522532127
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 555.529397922986
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0 0 0 0 0.148 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0 0 0.148 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 5. ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     LAIGV: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     OMD: 0.681908407324213 0.682295308592443 0.682647566048721 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.865630423728814 0.864377796610169 0.863211101694915 0. ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.865630423728814 0.864377796610169 0.863211101694915 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 47.939 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 3.99 7.78 11.31 16.04 20.55 23.86 27.42 31.52 36.14 40.8 ...
#>     STp: 3500.48
#>     WR: 130 130 130 130 130 129.790711380507 130 130 130 130 130 ...
#>     WRp: 111.684576106885
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 8726.33497883479
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 89 117 145 173 201 229 257 285
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 26
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2016
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.425137788528002 0.4583695833186 0.514875174072729 0.59 ...
#>     AgeDR: 502.41 502.7 502.7 505.98 510.39 512.89 512.89 514.12 51 ...
#>     AgeDRp: 3059.75216179612
#>     AgeDV: 502.41 502.7 502.7 505.98 510.39 512.89 512.89 514.12 51 ...
#>     AgeDVp: 386.16446805472
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2295.17765333274
#>     AgeGV: 102.41 102.7 102.7 105.98 110.39 112.89 112.89 114.12 11 ...
#>     AgeGVp: 370.305479244779
#>     BM: 747.7587 747.490907997 747.490907997 744.464666239853 74 ...
#>     BMDR: 29.9277 29.919020967 29.919020967 29.8208865782282 29.68 ...
#>     BMDRp: 0.500130124356253
#>     BMDV: 297.831 297.57188703 297.57188703 294.643779661625 290.7 ...
#>     BMDVp: 282.480008175896
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 140
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 491.438768295548
#>     BM_after_cut: 1070
#>     ENV: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     LAIGV: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     OMD: 0.677537527475466 0.677569768025088 0.677569768025088 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.858665597014925 0.858576865671642 0.858576865671642 0. ...
#>     OMDGR: 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0 ...
#>     OMDGV: 0.858665597014925 0.858576865671642 0.858576865671642 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.41 2.7 2.7 5.98 10.39 12.89 12.89 14.12 14.12 14.93 16 ...
#>     STp: 3005.38
#>     WR: 160 159.600913343104 159.873592083394 159.849885542053 1 ...
#>     WRp: 159.800708876818
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7844.84187034734
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 106 134 162 190 218 246 274 302
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 99
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: sorens1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2013
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[2]][[2]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.538462853575821 0.44108127367381 0.48117957128052 0.19 ...
#>     AgeDR: 502.03 505.18 509.1 512.19 513.89 515.61 522.11 528.58 5 ...
#>     AgeDRp: 2669.21142844763
#>     AgeDV: 502.03 505.18 509.1 512.19 513.89 515.61 522.11 528.58 5 ...
#>     AgeDVp: 418.359903859563
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2569.40156551628
#>     AgeGV: 102.03 105.18 109.1 112.19 113.89 115.61 122.11 128.58 1 ...
#>     AgeGVp: 413.555373379791
#>     BM: 748.1121 745.200056985 741.609687482843 738.812083485428 ...
#>     BMDR: 29.9391 29.844791835 29.7278002510068 29.6359413482312 2 ...
#>     BMDRp: 1.82722569752177
#>     BMDV: 298.173 295.35526515 291.881887231836 289.176142137197 2 ...
#>     BMDVp: 261.536447122351
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 140
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 731.411046472225
#>     BM_after_cut: 1070
#>     ENV: 0 0 0 0 0 0 0.3 0.294 0.562 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0.3 0.294 0.562 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     LAIGV: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     OMD: 0.677494554067618 0.677846661968646 0.67827836349193 0.6 ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.858781865671642 0.857818059701492 0.856618656716418 0. ...
#>     OMDGR: 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0 ...
#>     OMDGV: 0.858781865671642 0.857818059701492 0.856618656716418 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.03 5.18 9.1 12.19 13.89 15.61 22.11 28.58 36.39 40.79  ...
#>     STp: 3336.94
#>     WR: 160 160 160 160 160 159.392168239272 158.940698255806 15 ...
#>     WRp: 157.257794854547
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 10226.404797478
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 91 119 147 175 203 231 259 287
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 65
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: sorens1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2014
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[2]][[3]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.0043292133946835
#>     ABSDV: 2.13640254190842
#>     AET: 0.19920064755 0.199048204821215 0.553229998408238 0.5051 ...
#>     AgeDR: 500 500.15 504.87 508.03 508.03 508.03 509.26 513.06 519 ...
#>     AgeDRp: 2958.06678161877
#>     AgeDV: 500 500.15 504.87 508.03 508.03 508.03 509.26 513.06 519 ...
#>     AgeDVp: 411.077694631718
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2709.57058128289
#>     AgeGV: 100 100.15 104.87 108.03 108.03 108.03 109.26 113.06 119 ...
#>     AgeGVp: 367.30328845583
#>     BM: 750 749.8605 745.47283284 742.576027166913 742.576027166 ...
#>     BMDR: 30 29.9955 29.85392124 29.7595828488816 29.7595828488816 ...
#>     BMDRp: 0.730668927372743
#>     BMDV: 300 299.865 295.6189116 292.816444318032 292.81644431803 ...
#>     BMDVp: 270.430701507395
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 140
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 546.364532379602
#>     BM_after_cut: 1070
#>     ENV: 0 0 0 0 0 0 0 0 0.292 1 0 0 0.472 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0.292 1 0 0 0.472 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     LAIGV: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     OMD: 0.677265671641791 0.677282544702568 0.677816117294391 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.859402985074627 0.859357089552239 0.857912910447761 0. ...
#>     OMDGR: 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0 ...
#>     OMDGV: 0.859402985074627 0.859357089552239 0.857912910447761 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 0 0.15 4.87 8.03 8.03 8.03 9.26 13.06 19.52 30.44 33.27  ...
#>     STp: 3444.05
#>     WR: 159.80079935245 160 160 160 159.518930436167 159.3773082 ...
#>     WRp: 155.407170432611
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 8723.48037771063
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 118 146 174 202 230 258 286 314
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 97
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: sorens1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2015
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[2]][[4]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.288574824085632 0.248494987407072 0.272542889414208 0. ...
#>     AgeDR: 502.97 505.51 507.62 510.78 513.77 515.61 518.18 521.58  ...
#>     AgeDRp: 3294.61534088583
#>     AgeDV: 502.97 505.51 507.62 510.78 513.77 515.61 518.18 521.58  ...
#>     AgeDVp: 427.34599834697
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2492.27519480207
#>     AgeGV: 102.97 105.51 107.62 110.78 113.77 115.61 118.18 121.58  ...
#>     AgeGVp: 405.976159982019
#>     BM: 747.2379 744.896294574 742.965604418392 740.0920493681 7 ...
#>     BMDR: 29.9109 29.834926314 29.7719746194775 29.6778951796799 2 ...
#>     BMDRp: 0.362613930718022
#>     BMDV: 297.327 295.06136826 293.193629798914 290.41415418842 28 ...
#>     BMDVp: 262.620457016568
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 140
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 563.50299526363
#>     BM_after_cut: 1070
#>     ENV: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.11 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.11 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     LAIGV: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     OMD: 0.677600930797493 0.677883307425121 0.678114770221929 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.858494253731343 0.857717089552239 0.857071492537313 0. ...
#>     OMDGR: 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0 ...
#>     OMDGV: 0.858494253731343 0.857717089552239 0.857071492537313 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.97 5.51 7.62 10.78 13.77 15.61 18.18 21.58 26.08 29.66 ...
#>     STp: 3185.95
#>     WR: 160 160 160 160 160 160 160 160 160 160 160 160 160 159. ...
#>     WRp: 140.565995242443
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 8659.23602568173
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 117 145 173 201 229 257 285
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 84
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: sorens1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2016
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#>