Example results of a parameter scan

The function run_parameter_scan() can take a significant time to execute, as it typically requires a few dozen model evaluations or more. In order to still showcase what its output can look like, and to facilitate testing and giving examples in the documentation of tools that make use of the output of run_parameter_scan() (such as e.g. analyze_parameter_scan()), this example dataset is provided.

Usage

parameter_scan_example

Format

A list containing an entry for each supplied parameter set in param_values. Each entry is itself a list containing the following keys:

params

The parameter set that was used to run growR for this entry.

data

A list containing for each simulated year a ModvegeSite object which was run for the respective year and therefore carries the respective results.

Details

The input for the parameter scan that produced this output was:

- `eps = 2e-2`
)

See also