Plot Parameter Scan Results

Visualize the results of a parameter scan and allow interactive inspection of model performance in parameter space.

Usage

plot_parameter_scan(analyzed, variable = "dBM", interactive = TRUE)

Arguments

analyzed

List; Output of analyze_parameter_scan().

variable

Str; Name of variable in analyzed to visualize. Can be changed later with PscanPlotter$set_variable(). Allowed values are the keys in analyzed except for params and metrics.

interactive

boolean; Toggle between just creating a static plot (interactive = FALSE) or entering a small, interactive analysis setting (interactive = TRUE, default).

Value

A PscanPlotter object.

Details

Under the hood this function just creates a PscanPlotter object and calls its analyze method.

See also

Examples

if (FALSE) { # interactive()
# There needs to be data available with which the modle is to be compared.
# For this example, use data provided by the package.
path = system.file("extdata", package = "growR")
datafile = file.path(path, "posieux1.csv")

# Analyze example output of `run_parameter_scan()`.
results = analyze_parameter_scan(parameter_scan_example, datafile = datafile)
# The following plots the results.
psp = plot_parameter_scan(results, interactive = FALSE)

# The interactive session can still be entered later from the returned 
# PscanPlotter object
psp$analyze()
}