Parameter Scan

Run ModVege for a different sets of parameters.

Usage

run_parameter_scan(environment, param_values, force = FALSE, outfilename = "")

Arguments

environment

Either a ModvegeEnvironment instance with all the site, management and weather inputs expected by ModvegeSite$run() or a string representing the name of a config file to read in order to generate the ModvegeEnvironment with read_config(). Note that, in the latter case, only the first found configuration is used if there are more than one valid uncommented lines in the config file.

param_values

A named list where each key stands for a ModVege parameter, i.e. a member of ModvegeParameters$parameter_names. Each list entry then has to be a vector containing the allowed values for the respective parameter. All possible allowed combinations of these parameter values are then generated and fed into a ModVege run.

force

Boolean. By default (force = FALSE), the function first counts the number of parameter combinations that need to be run and asks the user, if it should proceed. This can be suppressed by letting force = TRUE.

outfilename

String. If nonempty, the results are stored as an rds file with filename outfilename using the saveRDS() function.

Value

results A list containing an entry for each supplied parameter set in param_values. Each entry is itself a list containing the following keys:

params

The parameter set that was used to run ModVege for this entry.

data

A list containing for each simulated year a ModvegeSite object which was run for the respective year and therefore carries the respective results.

Note

Special care has to be taken in the creation of the param_values argument. It's possible to choose values that do not allow for any valid combination. Confer create_combinations().

See also

Examples

env = create_example_environment()
#> [1] "==============================================================="
#> [1] "WHC"   "NI"    "w_FGA" "w_FGB" "w_FGC" "w_FGD"
#>  [1] "AgeGV"             "AgeGR"             "AgeDV"            
#>  [4] "AgeDR"             "BMGV"              "BMGR"             
#>  [7] "BMDV"              "BMDR"              "SENGV"            
#> [10] "SENGR"             "ABSDV"             "ABSDR"            
#> [13] "ST"                "cBM"               "LON"              
#> [16] "LAT"               "ELV"               "WHC"              
#> [19] "NI"                "RUEmax"            "w_FGA"            
#> [22] "w_FGB"             "w_FGC"             "w_FGD"            
#> [25] "sigmaGV"           "sigmaGR"           "T0"               
#> [28] "T1"                "T2"                "KGV"              
#> [31] "KGR"               "KlDV"              "KlDR"             
#> [34] "OMDDV"             "OMDDR"             "CO2_growth_factor"
#> [1] "==============================================================="
# We're creating a trivial list of parameters to explore here in order to 
# prevent the example from requiring a long time to execute. See 
# [create_combinations()] for more realistic uses of param_values.
param_values = list(w_FGA = c(0, 1), w_FGB = c(0, 1))
run_parameter_scan(env, param_values, force = TRUE)
#> [INFO]Parameter combination 1 out of 2.
#> [INFO]Starting run 1 out of 1.
#> [INFO][Run 1/1]Simulating year 2013 (1/4)
#> [INFO][Run 1/1]Simulating year 2014 (2/4)
#> [INFO][Run 1/1]Simulating year 2015 (3/4)
#> [INFO][Run 1/1]Simulating year 2016 (4/4)
#> [INFO]All runs completed.
#> [INFO]Parameter combination 2 out of 2.
#> [INFO]Starting run 1 out of 1.
#> [INFO][Run 1/1]Simulating year 2013 (1/4)
#> [INFO][Run 1/1]Simulating year 2014 (2/4)
#> [INFO][Run 1/1]Simulating year 2015 (3/4)
#> [INFO][Run 1/1]Simulating year 2016 (4/4)
#> [INFO]All runs completed.
#> [INFO]Completed parameter scan.
#> [INFO]Time used: 2.30351161956787
#> [[1]]
#> [[1]]$params
#> w_FGA w_FGB 
#>     0     1 
#> 
#> [[1]]$data
#> [[1]]$data[[1]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.000970254586886847
#>     ABSDV: 0.134223101067056
#>     AET: 0.457223659360304 0.492747302067495 0.480151022303239 0. ...
#>     AgeDR: 504.04 506.52 508.39 512.93 519.74 524.89 527.85 528.64  ...
#>     AgeDRp: 2837.58304961219
#>     AgeDV: 504.04 506.52 508.39 512.93 519.74 524.89 527.85 528.64  ...
#>     AgeDVp: 382.185972110443
#>     AgeGR: 0 2.48 4.35 8.89 15.7 20.85 23.81 24.6 24.6 26.91 30.53  ...
#>     AgeGRp: 2847.79257160506
#>     AgeGV: 104.04 106.52 108.39 112.93 119.74 124.89 127.85 128.64  ...
#>     AgeGVp: 460.027298698053
#>     BM: 777.4548 775.904723616 774.741574739753 771.927981545784 ...
#>     BMDR: 29.8788 29.804700576 29.7489657859229 29.6139054812548 2 ...
#>     BMDRp: 1.90245997428794
#>     BMDV: 297.576 296.10002304 294.99260895383 292.31407606453 288 ...
#>     BMDVp: 197.386913333905
#>     BMG: 450 450 450 450 450 450 450 450 450 450 450 450 450 450  ...
#>     BMGR: 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 473.585202806512
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0.362 0.0300000000000001 0 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0.362 0.0300000000000001 0 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.765 0.765 0.765 0.765 0.765 0.765 0.765 0.765 0.765 0. ...
#>     LAIGV: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     OMD: 0.687466872672212 0.687376254677945 0.687305943751135 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.863586 0.862603538461538 0.861862730769231 0.860064192 ...
#>     OMDGR: 0.9 0.898283076923077 0.896988461538462 0.89384538461538 ...
#>     OMDGV: 0.860985 0.860055 0.85935375 0.85765125 0.8550975 0.8531 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 4.04 6.52 8.39 12.93 19.74 24.89 27.85 28.64 28.64 30.95 ...
#>     STp: 3697.52
#>     WR: 130 129.513978768103 129.154466687398 128.690629344777 1 ...
#>     WRp: 129.595180612553
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 5509.03308489657
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 106 134 162 190 218 246 274 302
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 33
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (smooth_interval = 28) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2013
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> [[1]]$data[[2]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.572832822953001 0.48903850308452 0.497218890323204 0.2 ...
#>     AgeDR: 503.27 507.84 512.21 517.15 521.58 524.56 533.05 541.75  ...
#>     AgeDRp: 3609.16139469568
#>     AgeDV: 503.27 507.84 512.21 517.15 521.58 524.56 533.05 530.134 ...
#>     AgeDVp: 402.095996082907
#>     AgeGR: 0 4.57 8.94 13.88 18.31 21.29 29.78 38.48 46.95 54.07 59 ...
#>     AgeGRp: 3438.23890660655
#>     AgeGV: 103.27 107.84 112.21 117.15 121.58 124.56 133.05 138.894 ...
#>     AgeGVp: 502.119668422176
#>     BM: 777.9399 775.079180997 772.368063090232 769.329464387358 ...
#>     BMDR: 29.9019 29.765248317 29.6351741818547 29.4887764213963 2 ...
#>     BMDRp: 0.313795425238456
#>     BMDV: 298.038 295.31393268 292.732888908377 289.840687965962 2 ...
#>     BMDVp: 188.40903977099
#>     BMG: 450 450 450 450 450 450 450 452.590417199238 454.8067834 ...
#>     BMGR: 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 422.590417199238 424.8067834 ...
#>     BMGVp: 682.233496004573
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0.698 0.74 0.694 0.424 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0.698 0.74 0.694 0.424 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 8.6347239974615 7.38788747044971 4.1184102 ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.765 0.765 0.765 0.765 0.765 0.765 0.765 0.769403709238 ...
#>     LAIGV: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.718403709238 ...
#>     OMD: 0.68747320326416 0.6873073960817 0.687140164282573 0.686 ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.8638555 0.862045076923077 0.860313884615384 0.85835688 ...
#>     OMDGR: 0.9 0.896836153846154 0.893810769230769 0.89039076923076 ...
#>     OMDGV: 0.86127375 0.85956 0.85792125 0.85606875 0.8544075 0.853 ...
#>     PGRO: 0 0 0 0 0 0 0 23.8133590663583 21.7252469283353 19.82292 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 3.27 7.84 12.21 17.15 21.58 24.56 33.05 41.75 50.22 57.3 ...
#>     STp: 4148.57
#>     WR: 129.932408897425 130 130 130 129.695080572826 129.063098 ...
#>     WRp: 128.927030701673
#>     cBM: 0 0 0 0 0 0 0 1.08012440180491 1.42177474336492 1.421774 ...
#>     cBMp: 9175.11133061148
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 91 119 147 175 203 231 259 287
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 1.08012440180491 0.341650341560005 0 0 0 0 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 8
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (smooth_interval = 28) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2014
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> [[1]]$data[[3]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.00872560429146165
#>     ABSDV: 2.19473026516464
#>     AET: 0.19920064755 0.199007766069559 0.593319128727675 0.8132 ...
#>     AgeDR: 500 500 506.11 512.28 512.28 512.28 513.87 519.91 528.86 ...
#>     AgeDRp: 3217.78205400211
#>     AgeDV: 500 500 506.11 512.28 512.28 512.28 513.87 519.91 528.86 ...
#>     AgeDVp: 367.216899941366
#>     AgeGR: 0 0 6.11 12.28 12.28 12.28 13.87 19.91 28.86 42.56 48.63 ...
#>     AgeGRp: 3249.45289045185
#>     AgeGV: 100 100 106.11 112.28 112.28 112.28 113.87 119.91 128.86 ...
#>     AgeGVp: 419.661466255218
#>     BM: 780 780 776.1507 772.309969401 772.309969401 772.3099694 ...
#>     BMDR: 30 30 29.8167 29.632730961 29.632730961 29.632730961 29. ...
#>     BMDRp: 1.07128352258584
#>     BMDV: 300 300 296.334 292.67723844 292.67723844 292.67723844 2 ...
#>     BMDVp: 202.093026258254
#>     BMG: 450 450 450 450 450 450 450 450 450 454.787950195301 455 ...
#>     BMGR: 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 424.787950195301 425 ...
#>     BMGVp: 497.796370006886
#>     BM_after_cut: 900
#>     ENV: 0 0 0.222 0.234 0 0 0 0.208 0.79 1 0.214 0 0.324 0.464 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0.222 0.234 0 0 0 0.208 0.79 1 0.214 0 0.324 0.464 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 15.959833984336 1.59517865193112 0 6.3 ...
#>     GROGR: 0
#>     GROGV: 0.259622197390632
#>     LAI: 0.765 0.765 0.765 0.765 0.765 0.765 0.765 0.765 0.765 0. ...
#>     LAIGV: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     OMD: 0.6875 0.6875 0.68728631566009 0.687054064497056 0.68705 ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.865 0.865 0.8625795 0.860135230769231 0.86013523076923 ...
#>     OMDGR: 0.9 0.9 0.89577 0.891498461538462 0.891498461538462 0.89 ...
#>     OMDGV: 0.8625 0.8625 0.86020875 0.857895 0.857895 0.857895 0.85 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 32.571089763951 15.2124609186641 12.22 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0.181735538173442
#>     ST: 0 0 6.11 12.28 12.28 12.28 13.87 19.91 28.86 42.56 48.63 ...
#>     STp: 4116.88
#>     WR: 129.800802306981 130 130 129.657290839011 129.1781276950 ...
#>     WRp: 123.707919906768
#>     cBM: 0 0 0 0 0 0 0 0 0 3.33588388300948 3.33588388300948 3.33 ...
#>     cBMp: 6175.91112242251
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 90 118 146 174 202 230 258 286 314
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 3.33588388300948 0 0 0.735288221726704 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 10
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (smooth_interval = 28) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2015
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> [[1]]$data[[4]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.272542889414208 0.224447085399936 0.288574824085632 0. ...
#>     AgeDR: 505.19 510.18 514.91 520.84 526.55 531.06 535.82 541.12  ...
#>     AgeDRp: 3039.76183527887
#>     AgeDV: 505.19 510.18 514.91 520.84 526.55 531.06 535.82 540.891 ...
#>     AgeDVp: 386.650641168004
#>     AgeGR: 0 4.99 9.72 15.65 21.36 25.87 30.63 35.93 41.75 47.63 54 ...
#>     AgeGRp: 3059.01646962943
#>     AgeGV: 105.19 110.18 114.91 120.84 126.55 131.06 135.82 141.064 ...
#>     AgeGVp: 455.007468976343
#>     BM: 776.7303 773.618454663 770.697483214828 767.069271726514 ...
#>     BMDR: 29.8443 29.695376943 29.5549178100596 29.379657147446 29 ...
#>     BMDRp: 1.39894261079682
#>     BMDV: 296.886 293.92307772 291.142565404769 287.689614579068 2 ...
#>     BMDVp: 194.370398773874
#>     BMG: 450 450 450 450 450 450 450 450.049632388952 450.1879673 ...
#>     BMGR: 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420.049632388952 420.1879673 ...
#>     BMGVp: 555.459648524387
#>     BM_after_cut: 900
#>     ENV: 0.0380000000000001 0 0 0.186 0.142 0 0 0.06 0.164 0.176  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0.0380000000000001 0 0 0.186 0.142 0 0 0.06 0.164 0.176  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0.165441296506164 0.461116643714601 0.4949 ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.765 0.765 0.765 0.765 0.765 0.765 0.765 0.765084375061 ...
#>     LAIGV: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714084375061 ...
#>     OMD: 0.687457403168127 0.687272316327138 0.687086608125704 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.8631835 0.861206692307692 0.859332884615385 0.85698369 ...
#>     OMDGR: 0.9 0.896545384615385 0.893270769230769 0.88916538461538 ...
#>     OMDGV: 0.86055375 0.8586825 0.85690875 0.854685 0.85254375 0.85 ...
#>     PGRO: 0 0 0 0 0 12.0268782974336 5.62725498320287 5.6272549832 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 5.19 10.18 14.91 20.84 26.55 31.06 35.82 41.12 46.94 52. ...
#>     STp: 3903.56
#>     WR: 130 130 130 130 130 129.801714812682 130 130 130 130 130 ...
#>     WRp: 111.73277413641
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 6613.15482138648
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 89 117 145 173 201 229 257 285
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 6
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (smooth_interval = 28) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2016
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> 
#> 
#> [[2]]
#> [[2]]$params
#> w_FGA w_FGB 
#>     1     0 
#> 
#> [[2]]$data
#> [[2]]$data[[1]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.000145976959279232
#>     ABSDV: 0.146741785851066
#>     AET: 0.457223659360304 0.492747302067495 0.48030634474488 0.6 ...
#>     AgeDR: 504.04 506.52 508.39 512.93 519.74 524.89 527.85 528.64  ...
#>     AgeDRp: 3461.13107998501
#>     AgeDV: 504.04 506.52 508.39 512.93 519.74 524.89 527.85 528.64  ...
#>     AgeDVp: 314.738599135336
#>     AgeGR: 0 2.48 4.35 8.89 15.7 20.85 23.81 24.6 24.6 26.91 30.53  ...
#>     AgeGRp: 3082.47683899715
#>     AgeGV: 104.04 106.52 108.39 112.93 119.74 124.89 127.85 128.64  ...
#>     AgeGVp: 369.172421216372
#>     BM: 776.2428 773.963752416 772.2577853451 768.138754918036 7 ...
#>     BMDR: 29.8788 29.804700576 29.7489657859229 29.6139054812548 2 ...
#>     BMDRp: 0.286229331920062
#>     BMDV: 296.364 294.15905184 292.508819559178 288.524849436782 2 ...
#>     BMDVp: 215.796743898626
#>     BMG: 450 450 450 450 450 450 450 450 450 450 450 450 450 450  ...
#>     BMGR: 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 499.010487365081
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0.362 0.0300000000000001 0 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0.362 0.0300000000000001 0 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 1.0098 1.0098 1.0098 1.0098 1.0098 1.0098 1.0098 1.0098  ...
#>     LAIGV: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     OMD: 0.692059597847478 0.692333370762807 0.692536935548023 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.8708688 0.870105511111111 0.869529966666667 0.86813265 ...
#>     OMDGR: 0.9 0.898966666666667 0.8981875 0.896295833333333 0.8934 ...
#>     OMDGV: 0.868788 0.868044 0.867483 0.866121 0.864078 0.862533 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 4.04 6.52 8.39 12.93 19.74 24.89 27.85 28.64 28.64 30.95 ...
#>     STp: 3697.52
#>     WR: 130 129.513978768103 129.154311364956 128.690127006237 1 ...
#>     WRp: 129.595180612553
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7514.24767775912
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 106 134 162 190 218 246 274 302
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 33
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (smooth_interval = 28) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2013
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> [[2]]$data[[2]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.572832822953001 0.489060458879845 0.497218890323204 0. ...
#>     AgeDR: 503.27 507.84 512.21 517.15 521.58 524.56 533.05 541.75  ...
#>     AgeDRp: 3475.71768762852
#>     AgeDV: 503.27 507.84 512.21 517.15 521.58 524.56 533.05 527.177 ...
#>     AgeDVp: 330.448929791844
#>     AgeGR: 0 4.57 8.94 13.88 18.31 21.29 29.78 38.48 46.95 54.07 59 ...
#>     AgeGRp: 3475.71854700041
#>     AgeGV: 103.27 107.84 112.21 117.15 121.58 124.56 133.05 137.738 ...
#>     AgeGVp: 429.339958294922
#>     BM: 776.9589 772.749596847 768.778497902626 764.346996084626 ...
#>     BMDR: 29.9019 29.765248317 29.6351741818547 29.4887764213963 2 ...
#>     BMDRp: 0.706211976449096
#>     BMDV: 297.057 292.98434853 289.143323720772 284.85821966323 28 ...
#>     BMDVp: 235.001167904304
#>     BMG: 450 450 450 450 450 450 450 454.9241080224 458.230837442 ...
#>     BMGR: 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 424.9241080224 428.230837442 ...
#>     BMGVp: 826.498017144479
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0.698 0.74 0.694 0.424 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0.698 0.74 0.694 0.424 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 12.2321080223999 10.5049438101883 5.865144 ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 1.0098 1.0098 1.0098 1.0098 1.0098 1.0098 1.0098 1.02084 ...
#>     LAIGV: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     OMD: 0.691959883592298 0.692467641974387 0.692941278665371 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.8710844 0.869677855555556 0.868332866666667 0.86681244 ...
#>     OMDGR: 0.9 0.898095833333333 0.896275 0.894216666666667 0.89237 ...
#>     OMDGV: 0.869019 0.867648 0.866337 0.864855 0.863526 0.862632 0. ...
#>     PGRO: 0 0 0 0 0 0 0 29.5176351891889 27.0300118623618 24.70158 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 3.27 7.84 12.21 17.15 21.58 24.56 33.05 41.75 50.22 57.3 ...
#>     STp: 4148.57
#>     WR: 129.932408897425 130 130 130 129.695080572826 129.062968 ...
#>     WRp: 128.915076456522
#>     cBM: 0 0 0 0 0 0 0 1.97118533810194 2.52386301971659 2.523863 ...
#>     cBMp: 11438.8822020437
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 91 119 147 175 203 231 259 287
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 1.97118533810194 0.552677681614642 0 0 0 0 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 8
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (smooth_interval = 28) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2014
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> [[2]]$data[[3]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.000737867338936885
#>     ABSDV: 2.44239292012923
#>     AET: 0.19920064755 0.19903411414677 0.593319128727675 0.81323 ...
#>     AgeDR: 500 500 506.11 512.28 512.28 512.28 513.87 519.91 528.86 ...
#>     AgeDRp: 4261.84503582437
#>     AgeDV: 500 500 506.11 512.28 512.28 512.28 513.87 519.91 528.86 ...
#>     AgeDVp: 296.268635341899
#>     AgeGR: 0 0 6.11 12.28 12.28 12.28 13.87 19.91 28.86 42.56 48.63 ...
#>     AgeGRp: 3560.51094361825
#>     AgeGV: 100 100 106.11 112.28 112.28 112.28 113.87 119.91 128.86 ...
#>     AgeGVp: 336.588313647143
#>     BM: 780 780 774.3177 768.682517451 768.682517451 768.6825174 ...
#>     BMDR: 30 30 29.8167 29.632730961 29.632730961 29.632730961 29. ...
#>     BMDRp: 0.0905914473832884
#>     BMDV: 300 300 294.501 289.04978649 289.04978649 289.04978649 2 ...
#>     BMDVp: 224.89805894376
#>     BMG: 450 450 450 450 450 450 450 450 450 461.100992872657 461 ...
#>     BMGR: 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 431.100992872657 431 ...
#>     BMGVp: 539.809805060868
#>     BM_after_cut: 900
#>     ENV: 0 0 0.222 0.234 0 0 0 0.208 0.79 1 0.214 0 0.324 0.464 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0.222 0.234 0 0 0 0.208 0.79 1 0.214 0 0.324 0.464 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 22.6089928726568 2.28305441181017 0 9. ...
#>     GROGR: 0
#>     GROGV: 0.384939897629929
#>     LAI: 1.0098 1.0098 1.0098 1.0098 1.0098 1.0098 1.0098 1.0098  ...
#>     LAIGV: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     OMD: 0.691538461538462 0.691538461538462 0.692229939984583 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.872 0.872 0.870119477777778 0.868220488888889 0.868220 ...
#>     OMDGR: 0.9 0.9 0.897454166666667 0.894883333333333 0.8948833333 ...
#>     OMDGV: 0.87 0.87 0.868167 0.866316 0.866316 0.866316 0.865839 0 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 40.3732015583158 19.0508545711797 15.3 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0.26945792834095
#>     ST: 0 0 6.11 12.28 12.28 12.28 13.87 19.91 28.86 42.56 48.63 ...
#>     STp: 4116.88
#>     WR: 129.800802306981 130 130 129.657290839011 129.1780182219 ...
#>     WRp: 116.194251760917
#>     cBM: 0 0 0 0 0 0 0 0 0 6.3091855127144 6.3091855127144 6.3091 ...
#>     cBMp: 8130.81789407314
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 90 118 146 174 202 230 258 286 314
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 6.3091855127144 0 0 1.42382416399697 0 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 10
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (smooth_interval = 28) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2015
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> [[2]]$data[[4]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.272542889414208 0.224447085399936 0.288574824085632 0. ...
#>     AgeDR: 505.19 510.18 514.91 520.84 526.55 531.06 535.82 541.12  ...
#>     AgeDRp: 3597.22779624457
#>     AgeDV: 505.19 510.18 514.91 520.84 526.55 531.06 535.82 540.785 ...
#>     AgeDVp: 315.805891221738
#>     AgeGR: 0 4.99 9.72 15.65 21.36 25.87 30.63 35.93 41.75 47.63 54 ...
#>     AgeGRp: 3300.68085449836
#>     AgeGV: 105.19 110.18 114.91 120.84 126.55 131.06 135.82 141.041 ...
#>     AgeGVp: 373.693515007609
#>     BM: 775.1733 770.603301813 766.334859226154 761.057783405748 ...
#>     BMDR: 29.8443 29.695376943 29.5549178100596 29.379657147446 29 ...
#>     BMDRp: 0.272241815065869
#>     BMDV: 295.329 290.90792487 286.779941416095 281.678126258302 2 ...
#>     BMDVp: 227.805702719572
#>     BMG: 450 450 450 450 450 450 450 450.070310150417 450.2662838 ...
#>     BMGR: 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420.070310150417 420.2662838 ...
#>     BMGVp: 628.320256660748
#>     BM_after_cut: 900
#>     ENV: 0.0380000000000001 0 0 0.186 0.142 0 0 0.06 0.164 0.176  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0.0380000000000001 0 0 0.186 0.142 0 0 0.06 0.164 0.176  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0.23436716805589 0.653245768938754 0.70128 ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 1.0098 1.0098 1.0098 1.0098 1.0098 1.0098 1.0098 1.00995 ...
#>     LAIGV: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     OMD: 0.692208864779011 0.692758077616239 0.69326453620777 0.6 ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.8705468 0.869010988888889 0.8675552 0.865730077777778  ...
#>     OMDGR: 0.9 0.897920833333333 0.89595 0.893479166666667 0.8911 0 ...
#>     OMDGV: 0.868443 0.866946 0.865527 0.863748 0.862035 0.860682 0. ...
#>     PGRO: 0 0 0 0 0 14.907808892444 6.97521333499673 6.97521333499 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 5.19 10.18 14.91 20.84 26.55 31.06 35.82 41.12 46.94 52. ...
#>     STp: 3903.56
#>     WR: 130 130 130 130 130 129.801714812682 130 130 130 130 130 ...
#>     WRp: 111.533599163167
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 9031.42442413255
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 89 117 145 173 201 229 257 285
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 6
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (smooth_interval = 28) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2016
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> 
#>