Willmott Index

Willmott's index of model performance as described in Willmott (2012).

Usage

willmott(predicted, observed, ...)

Arguments

predicted

Vector containing the predictions y.

observed

Vector containing the observations z.

...

Scaling factor c in the denominator in the Willmott index. The originally proposed value of 2 should be fine.

Value

willmott Value between -1 and 1

Details

This index takes on values from -1 to 1, where values closer to 1 are generally indicating better model performance. Values close to -1 can either mean that the model predictions differ strongly from the observation, or that the observations show small variance (or both).

References

Willmott CJ, Robeson SM, Matsuura K (2012). “A Refined Index of Model Performance.” International Journal of Climatology, 32(13), 2088–2094. ISSN 1097-0088, doi:10.1002/joc.2419 , https://rmets.onlinelibrary.wiley.com/doi/abs/10.1002/joc.2419.

See also

Examples

predicted = c(21.5, 22.2, 19.1)
observed = c(20, 20, 20)
# The Willmott index "fails" in this case, as the variance in the 
# observation is 0.
willmott(predicted, observed)
#> [1] -1

# Try with more realistic observations
observed = c(20.5, 19.5, 20.0)
willmott(predicted, observed)
#> [1] -0.5652174