Combinator

Helps to find all possible combinations for a given set of values.

See also

Public fields

combinations

list Once run, holds all valid parameter combinations as named lists.

eps

float Numerical precision to require when checking the functional group weight sum criterion.

Methods

Public methods

Combinator$create_combinations()

Combinator$clone()

Method create_combinations()

Find possible combinations

Usage

Combinator$create_combinations(param_values)

Arguments

param_values

A list giving all options for the parameter values which are to be combined. As an example:

list(w_FGA = c(0, 0.5, 1), w_FGB = c(0, 0.5, 1), NI = c(0.5, 0.9))

This would generate the combinations

w_FGA	w_FGB	NI

0	1	0.5

0	1	0.9

0.5	0.5	0.5

0.5	0.5	0.9

1	0	0.5

1	0	0.9

eps

Precision to be used when checking if the sum citerion of the functional groups (w_FGA + w_FGB + w_FGC + w_FGD = 1) is fulfilled.

Returns

combinations A list containing vectors of parameter value combinations.

Method clone()

The objects of this class are cloneable with this method.

Usage

Combinator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.