Representation of a grassland plant population

A functional group is a representation of a grassland plant population with certain functional attributes.

It contains many plant parameters that are collected under the hood of functional groups. The class implements S3 style operator overloading such that one can do things like

mixed_functional_group = 0.8 * FG_A + 0.2 * FG_B

Public fields

SLA Specific Leaf Area in m^2^ per g.

pcLAM Percentage of laminae (number between 0 and 1).

ST1 Temperature sum in degree Celsiues days after which the seasonality function SEA starts to decrease from its maximum plateau. See also SEA().

ST2 Temperature sum in degree Celsiues days after which the seasonality function SEA has decreased back to its minimum value. See also SEA().

maxSEA Maximum value of the seasonality function SEA()

minSEA Minimum value of the seasonality function SEA(). Usually, minSEA = 1 - (maxSEA - 1).

maxOMDGV Maximum organic matter digestability for green vegetative matter in arbitrary units.

minOMDGV Minimum organic matter digestability for green vegetative matter in arbitrary units.

maxOMDGR Maximum organic matter digestability for green reproductive matter in arbitrary units.

minOMDGR Minimum organic matter digestability for green reproductive matter in arbitrary units.

BDGV Bulk density of green vegetative dry matter in g per m^3^.

BDDV Bulk density of dead vegetative dry matter in g per m^3^.

BDGR Bulk density of green reproductive dry matter in g per m^3^.

BDDR Bulk density of dead reproductive dry matter in g per m^3^.

fg_parameter_names Vector of strings of the variable names of all vegetation parameters governed by functional group composition.

Default values for parameters are taken from functional group A in Jouven et al.

References

Jouven M, Carrère P, Baumont R (2006). “Model Predicting Dynamics of Biomass, Structure and Digestibility of Herbage in Managed Permanent Pastures. 1. Model Description.” Grass and Forage Science, 61(2), 112--124. ISSN 1365-2494, doi:10.1111/j.1365-2494.2006.00515.x , https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1365-2494.2006.00515.x.

Public fields

fg_parameter_names

Names of the vegetation parameters governed by functional group composition.

Methods

Public methods

FunctionalGroup$new()

FunctionalGroup$get_parameters()

FunctionalGroup$get_parameters_ordered()

FunctionalGroup$set_parameters()

FunctionalGroup$set_parameters_ordered()

FunctionalGroup$clone()

Method new()

Constructor

Usage

FunctionalGroup$new(...)

Arguments

...

Key-value pairs of parameters to be set.

Method get_parameters()

Convenient getter

Returns all parameters with their names in a list.

Usage

FunctionalGroup$get_parameters()

Method get_parameters_ordered()

Ordered getter

Returns all parameters in reproducible order in a vector.

Usage

FunctionalGroup$get_parameters_ordered()

Method set_parameters()

Convenient setter

Set all specified parameters.

Usage

FunctionalGroup$set_parameters(...)

Arguments

...

Key-value pairs of parameters to be set.

Method set_parameters_ordered()

Efficient setter, assumes parameters come in known order.

Usage

FunctionalGroup$set_parameters_ordered(ordered_parameter_values)

Arguments

ordered_parameter_values

Parameter values to be set. Need to be in the same order as pFunctionalGroup]$fg_parameter_names.

Method clone()

The objects of this class are cloneable with this method.

Usage

FunctionalGroup$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.