Parameter Data Object

Data structure that contains site and vegetation parameters necessary for the configuration of an growR simulation run.

Note

Programmatically speaking, all parameters described under Parameter description are also fields of this R6Class.

Parameter description

The following is a list and description of model parameters, including the vegetation parameters, which are defined through the functional group composition.

Site and model parameters

LON geographic longitude of site in degree.

LAT geographic latitude of site in degree.

ELV geographic elevation of site in m.a.s.l.

WHC water-holding capacity of site in mm.

NI site nutritional index (dimensionless).

RUEmax maximum radiuation use efficiency in g DM per MJ.

w_FGA relative weight of functional group A.

w_FGB relative weight of functional group B.

w_FGC relative weight of functional group C.

w_FGD relative weight of functional group D.

sigmaGV rate of GV respirative biomass loss (dimensionless).

sigmaGR rate of GR respirative biomass loss (dimensionless).

T0 photosynthesis activation temperature (degree C).

T1 photosynthesis plateau temperature (degree C).

T2 photosynthesis max temperature (degree C).

KGV basic senescence rate GV (dimensionless).

KGR basic senescence rate GR (dimensionless).

KlGV basic abscission rate GV (dimensionless).

KlGR basic abscission rate GR (dimensionless).

maxOMDDV organic matter digestibility in gram per gram DV.

minOMDDR organic matter digestibility in gram per gram DR.

CO2_growth_factor strength of effect of CO2 concentration on growth. See parameter b in fCO2_growth_mod().

crop_coefficient multiplicative factor K~c~ by which reference evapotranspiration ET~0~ has to be multiplied to get the crop evapotranspiration ET~c~: ET~c~ = K~c~ ET~0~

senescence_cap fraction c~s~ of GRO to which SEN is limited: SEN~i~^max^ = c~s~ GRO~i~ for i in GV, GR. Makes it less likely for grass population to die out. Can be set to large values in order to effectively disable senescence capping.

stubble_height float. Minimum height the grass can assume. The biomass will not fall below that height. This effectively presents a simple model of plant reserves.

SGS_method string. Choice of method to determine the start of the growing season. Can be either "MTD" for the multicriterial thermal definition (see start_of_growing_season_mtd()) or "simple" for a commonly used approach as described in start_of_growing_season()).

Initial conditions

AgeGV Age of green vegetative matter in degree Celsius days.

AgeGR Age of green reproductive matter in degree Celsius days.

AgeDV Age of dead vegetative matter in degree Celsius days.

AgeDR Age of dead reproductive matter in degree Celsius days.

BMGV biomass of GV (kg DM per ha).

BMGR biomass of GR (kg DM per ha).

BMDV biomass of DV (kg DM per ha).

BMDR biomass of DR (kg DM per ha).

BMDR biomass of DR (kg DM per ha).

SENG senescence of GV (kg DM per ha).

SENG senescence of GR (kg DM per ha).

ABSG abscission of DV (kg DM per ha).

ABSG abscission of DR (kg DM per ha).

ST thermal time (degree days).

cBM cumulative total biomass (kg per ha).

Vegetation parameters

SLA Specific Leaf Area in m^2^ per g.

pcLAM Percentage of laminae (number between 0 and 1).

ST1 Temperature sum in degree Celsiues days after which the seasonality function SEA starts to decrease from its maximum plateau. See also SEA().

ST2 Temperature sum in degree Celsiues days after which the seasonality function SEA has decreased back to its minimum value. See also SEA().

maxSEA Maximum value of the seasonality function SEA()

minSEA Minimum value of the seasonality function SEA(). Usually, minSEA = 1 - (maxSEA - 1).

maxOMDGV Maximum organic matter digestability for green vegetative matter in arbitrary units.

minOMDGV Minimum organic matter digestability for green vegetative matter in arbitrary units.

maxOMDGR Maximum organic matter digestability for green reproductive matter in arbitrary units.

minOMDGR Minimum organic matter digestability for green reproductive matter in arbitrary units.

BDGV Bulk density of green vegetative dry matter in g per m^3^.

BDDV Bulk density of dead vegetative dry matter in g per m^3^.

BDGR Bulk density of green reproductive dry matter in g per m^3^.

BDDR Bulk density of dead reproductive dry matter in g per m^3^.

fg_parameter_names Vector of strings of the variable names of all vegetation parameters governed by functional group composition.

Public fields

required_parameter_names

Names of parameters that do not have a default value and are therefore strictly required.

parameter_names

Names of all required and optional parameters and state variables.

n_parameters

Number of total parameters.

functional_group

The FunctionalGroup instance holding the vegetation parameters.

fg_parameter_names

Names of vegetation parameters defined by the functional group composition.

initial_condition_names

Names of initial conditions.

param_file

Name of the parameter file from which initial parameter values were read.

Methods

Public methods

ModvegeParameters$new()

ModvegeParameters$read_parameters()

ModvegeParameters$set_parameters()

ModvegeParameters$update_functional_group()

ModvegeParameters$check_parameters()

ModvegeParameters$clone()

Method new()

Constructor

Usage

ModvegeParameters$new(param_file = NULL)

Arguments

param_file

Name of file containing the site and vegetation parameters.

Method read_parameters()

Read parameters from parameter file

Reads in parameters from the supplied param_file and stores them in internal fields.

This function carries out some basic sanity checks of the supplied param_file and reports on unidentified and missing parameter names.

Usage

ModvegeParameters$read_parameters(param_file)

Arguments

param_file

Path or name of file to read parameters from.

Returns

P List with field names as in the class variable parameter_names.

Method set_parameters()

Savely update the given parameters

This is the preferred method for changing the internal parameter values, because special care is taken to account for potential changes to functional group weights.

Usage

ModvegeParameters$set_parameters(params)

Arguments

params

List of name-value pairs of the parameters to update.

Method update_functional_group()

Update functional group parameters

Should be run whenever the functional group composition is changed in order to reflect the changes in the parameter list self$P.

Usage

ModvegeParameters$update_functional_group()

Method check_parameters()

Parameter Sanity Check Ensure that the supplied params are valid ModVege parameters and, if requested, check that all required parameters are present. Issues a warning for any invalid parameters and throws an error if completeness is not satisfied (only when check_for_completeness = TRUE).

Usage

ModvegeParameters$check_parameters(param_names, check_for_completeness = TRUE)

Arguments

param_names

A list of parameter names to be checked.

check_for_completeness

Boolean Toggle whether only the validity of supplied param_names is checked or whether we want to check that all required parameters to be present (default). In the latter case, if any required parameter is missing, an error is thrown.

Returns

not_known The list of unrecognized parameter names.

Method clone()

The objects of this class are cloneable with this method.

Usage

ModvegeParameters$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.