Run growR simulations

Start the loop over runs specified in the config file.

Usage

growR_run_loop(
  modvege_environments,
  write_files = TRUE,
  store_results = FALSE,
  output_dir = NULL
)

Arguments

modvege_environments

A list of ModvegeEnvironment instances.

write_files

boolean; If TRUE, write results of ModvegeSite$run() to files.

store_results

boolean; If TRUE, return a list of the ModvegeSite objects which were run.

output_dir

string; name of directory to which output files are to be written if write_files is TRUE.

Value

If store_results == TRUE, a list of the format [[run]][[year]] containing clones of the ModvegeSite instances that were run. Otherwise an empty list. Defaults to getOption("growR.output_dir").

Details

By default, returns an empty list but writes output to the output files as specified in the site_name and run_name fields of the supplied ModvegeEnvironment instances. Change this behaviour through the write_files and store_results arguments.

Examples

env1 = create_example_environment(site = "posieux")
env2 = create_example_environment(site = "sorens")

growR_run_loop(c(env1, env2), write_files = FALSE, 
store_results = TRUE)
#> [INFO]Starting run 1 out of 2.
#> [INFO][Run 1/2]Simulating year 2013 (1/4)
#> [INFO][Run 1/2]Simulating year 2014 (2/4)
#> [INFO][Run 1/2]Simulating year 2015 (3/4)
#> [INFO][Run 1/2]Simulating year 2016 (4/4)
#> [INFO]Starting run 2 out of 2.
#> [INFO][Run 2/2]Simulating year 2013 (1/4)
#> [INFO][Run 2/2]Simulating year 2014 (2/4)
#> [INFO][Run 2/2]Simulating year 2015 (3/4)
#> [INFO][Run 2/2]Simulating year 2016 (4/4)
#> [INFO]All runs completed.
#> [[1]]
#> [[1]][[1]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 9.2911798258096e-05
#>     ABSDV: 0.142929803932849
#>     AET: 0.457223659360304 0.492747302067495 0.48026211911084 0.6 ...
#>     AgeDR: 504.04 506.52 508.39 512.93 519.74 524.89 527.85 528.64  ...
#>     AgeDRp: 3710.24231919844
#>     AgeDV: 504.04 506.52 508.39 512.93 519.74 524.89 527.85 528.64  ...
#>     AgeDVp: 332.99019575139
#>     AgeGR: 0 2.48 4.35 8.89 15.7 20.85 23.81 24.6 24.6 26.91 30.53  ...
#>     AgeGRp: 3081.25048755388
#>     AgeGV: 104.04 106.52 108.39 112.93 119.74 124.89 127.85 128.64  ...
#>     AgeGVp: 391.890175821018
#>     BM: 776.2428 773.963752416 772.2577853451 768.138754918036 7 ...
#>     BMDR: 29.8788 29.804700576 29.7489657859229 29.6139054812548 2 ...
#>     BMDRp: 0.182179996584502
#>     BMDV: 296.364 294.15905184 292.508819559178 288.524849436782 2 ...
#>     BMDVp: 210.190888136543
#>     BMG: 450 450 450 450 450 450 450 450 450 450 450 450 450 450  ...
#>     BMGR: 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 492.002726504134
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0.362 0.0300000000000001 0 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0.362 0.0300000000000001 0 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.93636 0.93636 0.93636 0.93636 0.93636 0.93636 0.93636  ...
#>     LAIGV: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     OMD: 0.690342193561261 0.690561449566593 0.690723723419518 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.867906305084746 0.867057949382206 0.866418261816177 0. ...
#>     OMDGR: 0.9 0.898749918699187 0.897807317073171 0.89551886178861 ...
#>     OMDGV: 0.865613898305085 0.864794237288136 0.864176186440678 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 4.04 6.52 8.39 12.93 19.74 24.89 27.85 28.64 28.64 30.95 ...
#>     STp: 3697.52
#>     WR: 130 129.513978768103 129.15435559059 128.690270052271 12 ...
#>     WRp: 129.595180612553
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7280.91533108682
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 106 134 162 190 218 246 274 302
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 33
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function () 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2013
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> [[1]][[2]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.572832822953001 0.489054207310794 0.497218890323204 0. ...
#>     AgeDR: 503.27 507.84 512.21 517.15 521.58 524.56 533.05 541.75  ...
#>     AgeDRp: 3493.70854482793
#>     AgeDV: 503.27 507.84 512.21 517.15 521.58 524.56 533.05 527.177 ...
#>     AgeDVp: 347.833214290326
#>     AgeGR: 0 4.57 8.94 13.88 18.31 21.29 29.78 38.48 46.95 54.07 59 ...
#>     AgeGRp: 3459.9810027108
#>     AgeGV: 103.27 107.84 112.21 117.15 121.58 124.56 133.05 138.093 ...
#>     AgeGVp: 447.797549017822
#>     BM: 776.9589 772.749596847 768.778497902626 764.346996084626 ...
#>     BMDR: 29.9019 29.765248317 29.6351741818547 29.4887764213963 2 ...
#>     BMDRp: 0.700246724940568
#>     BMDV: 297.057 292.98434853 289.143323720772 284.85821966323 28 ...
#>     BMDVp: 221.574044405176
#>     BMG: 450 450 450 450 450 450 450 453.81366924764 456.67125541 ...
#>     BMGR: 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 423.81366924764 426.67125541 ...
#>     BMGVp: 789.616414833596
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0.698 0.74 0.694 0.424 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0.698 0.74 0.694 0.424 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 11.1216692476401 9.52528721338292 5.312077 ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.93636 0.93636 0.93636 0.93636 0.93636 0.93636 0.93636  ...
#>     LAIGV: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     OMD: 0.690256760980296 0.690663957365493 0.691040536462754 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.868143830508475 0.866580529879197 0.865085645032382 0. ...
#>     OMDGR: 0.9 0.897696422764228 0.895493658536585 0.89300357723577 ...
#>     OMDGV: 0.865868389830509 0.864357966101695 0.862913644067797 0. ...
#>     PGRO: 0 0 0 0 0 0 0 27.8836414973678 25.4641887083641 23.24394 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 3.27 7.84 12.21 17.15 21.58 24.56 33.05 41.75 50.22 57.3 ...
#>     STp: 4148.57
#>     WR: 129.932408897425 130 130 130 129.695080572826 129.063005 ...
#>     WRp: 128.918147962918
#>     cBM: 0 0 0 0 0 0 0 0.860746563342184 0.860746563342184 0.8607 ...
#>     cBMp: 11091.3123738518
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 91 119 147 175 203 231 259 287
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0.860746563342184 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 8
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function () 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2014
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> [[1]][[3]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.000619929903798892
#>     ABSDV: 2.36137853262158
#>     AET: 0.19920064755 0.199026611944203 0.593319128727675 0.8132 ...
#>     AgeDR: 500 500 506.11 512.28 512.28 512.28 513.87 519.91 528.86 ...
#>     AgeDRp: 4501.01548571705
#>     AgeDV: 500 500 506.11 512.28 512.28 512.28 513.87 519.91 528.86 ...
#>     AgeDVp: 315.730806052637
#>     AgeGR: 0 0 6.11 12.28 12.28 12.28 13.87 19.91 28.86 42.56 48.63 ...
#>     AgeGRp: 3617.4757021991
#>     AgeGV: 100 100 106.11 112.28 112.28 112.28 113.87 119.91 128.86 ...
#>     AgeGVp: 357.07105403138
#>     BM: 780 780 774.3177 768.682517451 768.682517451 768.6825174 ...
#>     BMDR: 30 30 29.8167 29.632730961 29.632730961 29.632730961 29. ...
#>     BMDRp: 0.0761117131735902
#>     BMDV: 300 300 294.501 289.04978649 289.04978649 289.04978649 2 ...
#>     BMDVp: 217.438170591305
#>     BMG: 450 450 450 450 450 450 450 450 450 459.028373963356 459 ...
#>     BMGR: 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 ...
#>     BMGRp: 123.141832042459
#>     BMGV: 420 420 420 420 420 420 420 420 420 429.028373963356 429 ...
#>     BMGVp: 528.833183487219
#>     BM_after_cut: 900
#>     ENV: 0 0 0.222 0.234 0 0 0 0.208 0.79 1 0.214 0 0.324 0.464 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0.222 0.234 0 0 0 0.208 0.79 1 0.214 0 0.324 0.464 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 20.5363739633564 2.06903170169447 0 8. ...
#>     GROGR: 0
#>     GROGV: 0.343288239209915
#>     LAI: 0.93636 0.93636 0.93636 0.93636 0.93636 0.93636 0.93636  ...
#>     LAIGV: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     OMD: 0.689895697522816 0.689895697522816 0.690453321806266 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.869152542372881 0.869152542372881 0.86706244021864 0.8 ...
#>     OMDGR: 0.9 0.9 0.896920162601626 0.893810081300813 0.8938100813 ...
#>     OMDGV: 0.866949152542373 0.866949152542373 0.864929745762712 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 38.1008793383235 17.937611201901 14.38 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0.24030176744694
#>     ST: 0 0 6.11 12.28 12.28 12.28 13.87 19.91 28.86 42.56 48.63 ...
#>     STp: 4116.88
#>     WR: 129.800802306981 130 130 129.657290839011 129.1780493927 ...
#>     WRp: 119.077370593625
#>     cBM: 0 0 0 0 0 0 0 0 0 4.23656660341395 4.23656660341395 4.23 ...
#>     cBMp: 7862.97939789055
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 90 118 146 174 202 230 258 286 314
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 4.23656660341395 0 0 0.575706351970084 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 10
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function () 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2015
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> [[1]][[4]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.272542889414208 0.224447085399936 0.288574824085632 0. ...
#>     AgeDR: 505.19 510.18 514.91 520.84 526.55 531.06 535.82 541.12  ...
#>     AgeDRp: 3819.67730203885
#>     AgeDV: 505.19 510.18 514.91 520.84 526.55 531.06 535.82 540.814 ...
#>     AgeDVp: 332.234387405887
#>     AgeGR: 0 4.99 9.72 15.65 21.36 25.87 30.63 35.93 41.75 47.63 54 ...
#>     AgeGRp: 3304.73869659475
#>     AgeGV: 105.19 110.18 114.91 120.84 126.55 131.06 135.82 141.048 ...
#>     AgeGVp: 393.9800012962
#>     BM: 775.1733 770.603301813 766.334859226154 761.057783405748 ...
#>     BMDR: 29.8443 29.695376943 29.5549178100596 29.379657147446 29 ...
#>     BMDRp: 0.161793317427512
#>     BMDV: 295.329 290.90792487 286.779941416095 281.678126258302 2 ...
#>     BMDVp: 216.419494233213
#>     BMG: 450 450 450 450 450 450 450 450.0641448248 450.241025208 ...
#>     BMGR: 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420.0641448248 420.241025208 ...
#>     BMGVp: 610.042278795913
#>     BM_after_cut: 900
#>     ENV: 0.0380000000000001 0 0 0.186 0.142 0 0 0.06 0.164 0.176  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0.0380000000000001 0 0 0.186 0.142 0 0 0.06 0.164 0.176  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0.21381608266647 0.589601279877797 0.63842 ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.93636 0.93636 0.93636 0.93636 0.93636 0.93636 0.93636  ...
#>     LAIGV: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     OMD: 0.690470081586808 0.690909030902664 0.691309918784962 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.867551559322034 0.865844585549584 0.864226552294337 0. ...
#>     OMDGR: 0.9 0.897484715447154 0.895100487804878 0.89211138211382 ...
#>     OMDGV: 0.865233813559322 0.863584576271186 0.862021271186441 0. ...
#>     PGRO: 0 0 0 0 0 14.062566911735 6.55903307450325 6.61150533909 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 5.19 10.18 14.91 20.84 26.55 31.06 35.82 41.12 46.94 52. ...
#>     STp: 3903.56
#>     WR: 130 130 130 130 130 129.801714812682 130 130 130 130 130 ...
#>     WRp: 111.587219003766
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 8763.73349308525
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 89 117 145 173 201 229 257 285
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 6
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function () 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2016
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.425137788528002 0.4583695833186 0.515292206584979 0.59 ...
#>     AgeDR: 503.61 505.1 505.96 510.44 516.05 519.75 520.81 523.24 5 ...
#>     AgeDRp: 2657.32101804528
#>     AgeDV: 503.61 505.1 505.96 510.44 516.05 519.75 520.81 523.24 5 ...
#>     AgeDVp: 402.971516618328
#>     AgeGR: 0 1.49 2.35 6.83 12.44 16.14 17.2 19.63 20.83 22.84 25.2 ...
#>     AgeGRp: 2611.14747753357
#>     AgeGV: 103.61 105.1 105.96 110.44 116.05 119.75 120.81 123.24 1 ...
#>     AgeGVp: 409.59851513876
#>     BM: 774.6427 773.271684397 772.483820568807 768.389958595924 ...
#>     BMDR: 29.8917 29.847161367 29.8214928082244 29.6878925204435 2 ...
#>     BMDRp: 2.25091345308294
#>     BMDV: 296.751 295.42452303 294.662327760583 290.70206607548 28 ...
#>     BMDVp: 269.208204690527
#>     BMG: 448 448 448 448 448 448 448 448 448 448 448 448 448 448  ...
#>     BMGR: 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 ...
#>     BMGRp: 140
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 502.155806916888
#>     BM_after_cut: 1070
#>     ENV: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.8651776 0.8651776 0.8651776 0.8651776 0.8651776 0.8651 ...
#>     LAIGV: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     OMD: 0.685348189289133 0.685491836840731 0.685573921772576 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.860279780783582 0.859802516526619 0.859527048566224 0. ...
#>     OMDGR: 0.89 0.889202204724409 0.888741732283465 0.8863429921259 ...
#>     OMDGV: 0.858298432835821 0.857842537313433 0.857579402985075 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 3.61 5.1 5.96 10.44 16.05 19.75 20.81 23.24 24.44 26.45  ...
#>     STp: 3364.63
#>     WR: 160 159.805206625861 159.94517316524 159.850342649062 15 ...
#>     WRp: 159.696976217117
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7236.91352648116
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 106 134 162 190 218 246 274 302
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 69
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function () 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: sorens1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2013
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> [[2]][[2]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.538462853575821 0.441078428207549 0.48117957128052 0.2 ...
#>     AgeDR: 503.23 507.58 512.7 516.99 519.89 522.81 530.51 538.18 5 ...
#>     AgeDRp: 4176.0618526471
#>     AgeDV: 503.23 507.58 512.7 516.99 519.89 522.81 530.51 538.18 5 ...
#>     AgeDVp: 400.920233490743
#>     AgeGR: 0 4.35 9.47 13.76 16.66 19.58 27.28 34.95 43.96 49.56 55 ...
#>     AgeGRp: 3410.87934110045
#>     AgeGV: 103.23 107.58 112.7 116.99 119.89 122.81 130.51 138.18 1 ...
#>     AgeGVp: 434.209909998041
#>     BM: 774.9961 770.988957865 766.332723212507 762.489925672435 ...
#>     BMDR: 29.9031 29.773021515 29.6205836448432 29.4935113410068 2 ...
#>     BMDRp: 0.130746560791537
#>     BMDV: 297.093 293.21593635 288.712139567664 284.996414331428 2 ...
#>     BMDVp: 266.380291224626
#>     BMG: 448 448 448 448 448 448 448 448 451.559355833998 451.813 ...
#>     BMGR: 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 ...
#>     BMGRp: 32.8153339444765
#>     BMGV: 420 420 420 420 420 420 420 420 423.559355833998 423.813 ...
#>     BMGVp: 805.481309858934
#>     BM_after_cut: 1070
#>     ENV: 0 0 0.024 0 0 0 0.54 0.534 0.802 0.12 0.092 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0.024 0 0 0 0.54 0.534 0.802 0.12 0.092 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 11.1277558339982 0.847699455764205 1.180 ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.8651776 0.8651776 0.8651776 0.8651776 0.8651776 0.8651 ...
#>     LAIGV: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     OMD: 0.685303144915008 0.685724906934686 0.686208371190375 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.860388782649254 0.858995427268187 0.857355431969092 0. ...
#>     OMDGR: 0.89 0.887670866141732 0.884929448818898 0.8826324409448 ...
#>     OMDGV: 0.858414701492537 0.857083731343284 0.855517164179105 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 27.1526561141134 13.8241920379029 25.110 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 3.23 7.58 12.7 16.99 19.89 22.81 30.51 38.18 47.19 52.79 ...
#>     STp: 3758.78
#>     WR: 159.9982653977 160 160 160 160 159.384049915908 158.9287 ...
#>     WRp: 158.403086751853
#>     cBM: 0 0 0 0 0 0 0 0 0.615718728023858 0.615718728023858 0.61 ...
#>     cBMp: 10463.7338845279
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 91 119 147 175 203 231 259 287
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0.615718728023858 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 9
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function () 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: sorens1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2014
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> [[2]][[3]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.0125393136776439
#>     ABSDV: 2.84294857675102
#>     AET: 0.19920064755 0.199053071325796 0.553229998408238 0.5051 ...
#>     AgeDR: 500 501.35 507.27 511.63 511.63 512.72 515.15 520.15 527 ...
#>     AgeDRp: 3046.67816107112
#>     AgeDV: 500 501.35 507.27 511.63 511.63 512.72 515.15 520.15 527 ...
#>     AgeDVp: 384.911533355196
#>     AgeGR: 0 1.35 7.27 11.63 11.63 12.72 15.15 20.15 27.81 39.93 43 ...
#>     AgeGRp: 3052.27342724932
#>     AgeGV: 100 101.35 107.27 111.63 111.63 112.72 115.15 120.15 127 ...
#>     AgeGVp: 370.945138142716
#>     BM: 778 776.7445 771.26071816 767.292168225174 767.292168225 ...
#>     BMDR: 30 29.9595 29.78213976 29.6522896306464 29.6522896306464 ...
#>     BMDRp: 1.62321212655585
#>     BMDV: 300 298.785 293.4785784 289.639878594528 289.63987859452 ...
#>     BMDVp: 276.014424927283
#>     BMG: 448 448 448 448 448 448 448 448 448 448 448 448 456.8837 ...
#>     BMGR: 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 ...
#>     BMGRp: 140
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 428.8837 ...
#>     BMGVp: 577.345225721953
#>     BM_after_cut: 1070
#>     ENV: 0 0 0.184 0 0 0 0 0 0.532 1 0 0 0.712 0.19 0.03599999999 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0.184 0 0 0 0 0 0.532 1 0 0 0.712 0.19 0.03599999999 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 16.0741257264739 0.7859907111905 ...
#>     GROGR: 0
#>     GROGV: 0.115354746321481
#>     LAI: 0.8651776 0.8651776 0.8651776 0.8651776 0.8651776 0.8651 ...
#>     LAIGV: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     OMD: 0.684921919963166 0.685054840152248 0.685636148003592 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.861315298507463 0.860882877871959 0.85898663330738 0.8 ...
#>     OMDGR: 0.89 0.889277165354331 0.886107401574803 0.8837729133858 ...
#>     OMDGV: 0.859402985074627 0.858989925373134 0.857178582089552 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 44.1800768664491 8.0954857471471 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0.0807483224250364
#>     ST: 0 1.35 7.27 11.63 11.63 12.72 15.15 20.15 27.81 39.93 43 ...
#>     STp: 3839.43
#>     WR: 159.80079935245 160 160 160 159.600132865521 159.0282826 ...
#>     WRp: 155.446655828424
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 6.26332303827531 6.2633230382753 ...
#>     cBMp: 8417.67896978055
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 118 146 174 202 230 258 286 314
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 6.26332303827531 0 0 0 0 0 0 0 0 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 13
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function () 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: sorens1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2015
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#> [[2]][[4]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.288574824085632 0.248494987407072 0.272542889414208 0. ...
#>     AgeDR: 504.17 507.91 511.22 515.58 519.77 522.81 526.58 531.18  ...
#>     AgeDRp: 2889.38267442745
#>     AgeDV: 504.17 507.91 511.22 515.58 519.77 522.81 526.58 531.18  ...
#>     AgeDVp: 430.689487882995
#>     AgeGR: 0 3.74 7.05 11.41 15.6 18.64 22.41 27.01 32.71 37.49 43. ...
#>     AgeGRp: 2815.87645682928
#>     AgeGV: 104.17 107.91 111.22 115.58 119.77 122.81 126.58 131.18  ...
#>     AgeGVp: 424.164182168505
#>     BM: 774.1219 770.686276534 767.679033979343 763.756307978488 ...
#>     BMDR: 29.8749 29.763167874 29.6646517883371 29.5353139065399 2 ...
#>     BMDRp: 1.37260291257627
#>     BMDV: 296.247 292.92310866 290.014382191006 286.220994071948 2 ...
#>     BMDVp: 260.309026640054
#>     BMG: 448 448 448 448 448 448 448 448 448 448 448 448 448 448  ...
#>     BMGR: 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 ...
#>     BMGRp: 140
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 622.067606067596
#>     BM_after_cut: 1070
#>     ENV: 0 0 0 0 0 0 0 0 0.14 0 0.114 0 0 0 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0.14 0 0.114 0 0 0 0 0 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.8651776 0.8651776 0.8651776 0.8651776 0.8651776 0.8651 ...
#>     LAIGV: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     OMD: 0.685414645434963 0.685774966142913 0.686086261185128 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.860119146455224 0.858921181139088 0.85786094980315 0.8 ...
#>     OMDGR: 0.89 0.887997480314961 0.886225196850394 0.8838907086614 ...
#>     OMDGV: 0.858127089552239 0.85698276119403 0.85597 0.85463597014 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 4.17 7.91 11.22 15.58 19.77 22.81 26.58 31.18 36.88 41.6 ...
#>     STp: 3584.19
#>     WR: 160 160 160 160 160 160 160 160 160 160 160 160 160 159. ...
#>     WRp: 139.995976157764
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 8439.81434686559
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 117 145 173 201 229 257 285
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 29
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function () 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: sorens1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2016
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#> 
#>