Run growR simulations

Start the loop over runs specified in the config file.

Usage

growR_run_loop(modvege_environments, output_dir = "", independent = TRUE)

Arguments

modvege_environments

A list of ModvegeEnvironment instances.

output_dir

string; name of directory to which output files are to be written. If output_dir == "" (default), no files are written.

independent

boolean; If TRUE (default) the simulation for each year starts with the same initial conditions, as specified in the parameters of the modvege_environments. If FALSE, initial conditions are taken as the final state values of the simulation of the previous year.

Value

A list of the format [[run]][[year]] containing clones of the ModvegeSite instances that were run. Also write to files, if output_dir is nonempty.

Details

By default, returns an empty list but writes output to the output files as specified in the site_name and run_name fields of the supplied ModvegeEnvironment instances. Change this behaviour through the write_files and store_results arguments.

Examples

env1 = create_example_environment(site = "posieux")
env2 = create_example_environment(site = "sorens")

growR_run_loop(c(env1, env2), output_dir = "")
#> [INFO]Starting run 1 out of 2.
#> [INFO] site name: `posieux1`.
#> [INFO]  run name: `-`.
#> [INFO][Run 1/2]Simulating year 2013 (1/4)
#> [INFO][Run 1/2]Simulating year 2014 (2/4)
#> [INFO][Run 1/2]Simulating year 2015 (3/4)
#> [INFO][Run 1/2]Simulating year 2016 (4/4)
#> [INFO]Starting run 2 out of 2.
#> [INFO] site name: `sorens1`.
#> [INFO]  run name: `-`.
#> [INFO][Run 2/2]Simulating year 2013 (1/4)
#> [INFO][Run 2/2]Simulating year 2014 (2/4)
#> [INFO][Run 2/2]Simulating year 2015 (3/4)
#> [INFO][Run 2/2]Simulating year 2016 (4/4)
#> [INFO]All runs completed.
#> [[1]]
#> [[1]][[1]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.457223659360304 0.492747302067495 0.48021377988667 0.6 ...
#>     AgeDR: 502.84 504.12 504.79 508.13 513.74 517.69 519.45 519.45  ...
#>     AgeDRp: 2605.87337482542
#>     AgeDV: 502.84 504.12 504.79 508.13 513.74 517.69 519.45 519.45  ...
#>     AgeDVp: 310.00327356826
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2563.53685693821
#>     AgeGV: 102.84 104.12 104.79 108.13 113.74 117.69 119.45 119.45  ...
#>     AgeGVp: 356.080235356464
#>     Autocut: NULL
#>     BM: 747.3162 746.116660416 745.491129848167 742.378930983185 ...
#>     BMDR: 29.8722 29.814845376 29.7848814563971 29.6356592003006 2 ...
#>     BMDRp: 2.34299102132323
#>     BMDV: 297.444 296.30181504 295.70624839177 292.743271782884 28 ...
#>     BMDVp: 219.862858385159
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 470.995494179339
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     LAIGV: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     OMD: 0.681803891792152 0.681942267363527 0.682014139566022 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.866010508474576 0.865587457627119 0.865366016949152 0. ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.866010508474576 0.865587457627119 0.865366016949152 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.84 4.12 4.79 8.13 13.74 17.69 19.45 19.45 19.45 20.56  ...
#>     WR: 130 129.509787758099 129.068168429521 128.689683936513 1 ...
#>     WRp: 129.627111149557
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7229.0639116253
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 106 134 162 190 218 246 274 302
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: -2.6838 -1.199539584 -0.62553056783328 -3.11219886498208 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 99
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2013
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 60
#>     minBMGV: 170
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[1]][[2]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.572832822953001 0.489080547324631 0.497218890323204 0. ...
#>     AgeDR: 502.07 505.44 508.61 512.35 515.58 517.36 524.65 532.15  ...
#>     AgeDRp: 3947.27747887845
#>     AgeDV: 502.07 505.44 508.61 512.35 515.58 517.36 524.65 532.15  ...
#>     AgeDVp: 367.259877424207
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 3070.87366438124
#>     AgeGV: 102.07 105.44 108.61 112.35 115.58 117.36 124.65 132.15  ...
#>     AgeGVp: 435.123915843282
#>     Autocut: NULL
#>     BM: 748.04385 744.87850580325 741.930399428063 738.484475911 ...
#>     BMDR: 29.90685 29.75567087325 29.6141826582477 29.448047093534 ...
#>     BMDRp: 0.145872941277318
#>     BMDV: 298.137 295.12283493 292.316216769816 289.036428817658 2 ...
#>     BMDVp: 202.749087689722
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 135.271106713856
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 707.307063935175
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0.458 0.5 0.454 0.184 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0.458 0.5 0.454 0.184 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     LAIGV: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     OMD: 0.681718979442181 0.682085787701713 0.68242443000321 0.6 ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.866265 0.865151186440678 0.864103474576271 0.862867372 ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.866265 0.865151186440678 0.864103474576271 0.862867372 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.07 5.44 8.61 12.35 15.58 17.36 24.65 32.15 39.42 45.34 ...
#>     WR: 129.946718652693 130 130 130 130 129.365338865 128.74430 ...
#>     WRp: 127.011229391066
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 11464.6810906814
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 91 119 147 175 203 231 259 287
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: -1.95615 -3.16534419675 -2.9481063751866 -3.445923516870 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 67
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2014
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 60
#>     minBMGV: 170
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[1]][[3]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.00897994410067624
#>     ABSDV: 1.83842313064963
#>     AET: 0.19920064755 0.19901996769842 0.592248295151082 0.81323 ...
#>     AgeDR: 500 500 504.91 509.88 509.88 509.88 510.27 515.11 522.86 ...
#>     AgeDRp: 2947.03539982489
#>     AgeDV: 500 500 504.91 509.88 509.88 509.88 510.27 515.11 522.86 ...
#>     AgeDVp: 320.820532488754
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2951.13482002105
#>     AgeGV: 100 100 104.91 109.88 109.88 109.88 110.27 115.11 122.86 ...
#>     AgeGVp: 338.246964126842
#>     Autocut: NULL
#>     BM: 750 750 745.36005 740.73093447225 740.73093447225 740.73 ...
#>     BMDR: 30 30 29.77905 29.55704718225 29.55704718225 29.55704718 ...
#>     BMDRp: 1.41527881807348
#>     BMDV: 300 300 295.581 291.17388729 291.17388729 291.17388729 2 ...
#>     BMDVp: 217.307698658349
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 509.617113025523
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0 0 0.55 1 0 0 0.084 0.224 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0.55 1 0 0 0.084 0.224 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     LAIGV: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     OMD: 0.681491525423729 0.681491525423729 0.682032984584464 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.866949152542373 0.866949152542373 0.865326355932203 0. ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.866949152542373 0.866949152542373 0.865326355932203 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 0 0 4.91 9.88 9.88 9.88 10.27 15.11 22.86 35.36 40.23 42 ...
#>     WR: 129.800799840891 129.603626755415 130 129.709778732205 1 ...
#>     WRp: 117.9543167997
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7395.53341352108
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 90 118 146 174 202 230 258 286 314
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 -4.63995 -4.62911552775 0 0 -0.357964320730916 -4.43 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 75
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2015
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 60
#>     minBMGV: 170
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[1]][[4]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.272542889414208 0.224447085399936 0.288574824085632 0. ...
#>     AgeDR: 503.99 507.78 511.31 516.04 520.55 523.86 527.42 531.52  ...
#>     AgeDRp: 2794.74655018318
#>     AgeDV: 503.99 507.78 511.31 516.04 520.55 523.86 527.42 531.52  ...
#>     AgeDVp: 339.499467337221
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2754.76329664012
#>     AgeGV: 103.99 107.78 111.31 116.04 120.55 123.86 127.42 131.52  ...
#>     AgeGVp: 383.051716604792
#>     Autocut: NULL
#>     BM: 746.22945 742.68975041175 739.429467580217 735.106022791 ...
#>     BMDR: 29.82045 29.65092074175 29.4939191164224 29.284659760291 ...
#>     BMDRp: 1.78392649662687
#>     BMDV: 296.409 293.03882967 289.935548463795 285.821363031093 2 ...
#>     BMDVp: 209.23252234107
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 555.529407167563
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0 0 0 0 0.148 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0 0 0.148 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     LAIGV: 0.873936 0.873936 0.873936 0.873936 0.873936 0.873936 0. ...
#>     OMD: 0.681931017284431 0.682339450549193 0.682711805926744 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.865630423728814 0.864377796610169 0.863211101694915 0. ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.865630423728814 0.864377796610169 0.863211101694915 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 3.99 7.78 11.31 16.04 20.55 23.86 27.42 31.52 36.14 40.8 ...
#>     WR: 130 130 130 130 130 129.790711380507 130 130 130 130 130 ...
#>     WRp: 111.684576099845
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 8655.18388124217
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 89 117 145 173 201 229 257 285
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: -3.77055 -3.53969958825 -3.26028283153287 -4.32344478883 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 85
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2016
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 60
#>     minBMGV: 170
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.425137788528002 0.4583695833186 0.514875174072729 0.59 ...
#>     AgeDR: 502.41 502.7 502.7 505.98 510.39 512.89 512.89 514.12 51 ...
#>     AgeDRp: 3040.04026390736
#>     AgeDV: 502.41 502.7 502.7 505.98 510.39 512.89 512.89 514.12 51 ...
#>     AgeDVp: 386.161133811613
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2295.17765333274
#>     AgeGV: 102.41 102.7 102.7 105.98 110.39 112.89 112.89 114.12 11 ...
#>     AgeGVp: 370.304808951
#>     Autocut: NULL
#>     BM: 747.72255 747.45043420575 747.45043420575 744.3753243852 ...
#>     BMDR: 29.89155 29.87854717575 29.87854717575 29.7315447236453  ...
#>     BMDRp: 0.47395457909275
#>     BMDV: 297.831 297.57188703 297.57188703 294.643779661625 290.7 ...
#>     BMDVp: 282.479780549091
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 140
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 491.438808210511
#>     BM_after_cut: 1070
#>     ENV: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     LAIGV: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     OMD: 0.677550945529553 0.677584798186734 0.677584798186734 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.858665597014925 0.858576865671642 0.858576865671642 0. ...
#>     OMDGR: 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0 ...
#>     OMDGV: 0.858665597014925 0.858576865671642 0.858576865671642 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.41 2.7 2.7 5.98 10.39 12.89 12.89 14.12 14.12 14.93 16 ...
#>     WR: 160 159.600913343104 159.873592083394 159.849885542053 1 ...
#>     WRp: 159.800708876818
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7841.72955659756
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 106 134 162 190 218 246 274 302
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: -2.27745 -0.27211579425 0 -3.07510982047989 -4.094811373 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 100
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: sorens1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2013
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 56
#>     minBMGV: 184
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[2]][[2]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.538462853575821 0.44108127367381 0.48117957128052 0.19 ...
#>     AgeDR: 502.03 505.18 509.1 512.19 513.89 515.61 522.11 528.58 5 ...
#>     AgeDRp: 2659.82013547961
#>     AgeDV: 502.03 505.18 509.1 512.19 513.89 515.61 522.11 528.58 5 ...
#>     AgeDVp: 418.362460982222
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2569.40156551628
#>     AgeGV: 102.03 105.18 109.1 112.19 113.89 115.61 122.11 128.58 1 ...
#>     AgeGVp: 413.555372388403
#>     Autocut: NULL
#>     BM: 748.08165 745.12259677875 741.474186950609 738.631281546 ...
#>     BMDR: 29.90865 29.76733162875 29.5922997187729 29.455139409576 ...
#>     BMDRp: 1.6535647814618
#>     BMDV: 298.173 295.35526515 291.881887231836 289.176142137197 2 ...
#>     BMDVp: 261.536446867659
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 140
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 731.411047687599
#>     BM_after_cut: 1070
#>     ENV: 0 0 0 0 0 0 0.3 0.294 0.562 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0.3 0.294 0.562 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     LAIGV: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     OMD: 0.677505849237299 0.677875545886319 0.678329217408393 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.858781865671642 0.857818059701492 0.856618656716418 0. ...
#>     OMDGR: 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0 ...
#>     OMDGV: 0.858781865671642 0.857818059701492 0.856618656716418 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.03 5.18 9.1 12.19 13.89 15.61 22.11 28.58 36.39 40.79  ...
#>     WR: 160 160 160 160 160 159.392168239272 158.940698255806 15 ...
#>     WRp: 157.257794854518
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 10217.0758797703
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 91 119 147 175 203 231 259 287
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: -1.91835 -2.95905322125 -3.64840982814105 -2.84290540383 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 67
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: sorens1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2014
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 56
#>     minBMGV: 184
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[2]][[3]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.00382810287884465
#>     ABSDV: 2.13640254190842
#>     AET: 0.19920064755 0.199048204821215 0.553229998408238 0.5051 ...
#>     AgeDR: 500 500.15 504.87 508.03 508.03 508.03 509.26 513.06 519 ...
#>     AgeDRp: 2941.37781938486
#>     AgeDV: 500 500.15 504.87 508.03 508.03 508.03 509.26 513.06 519 ...
#>     AgeDVp: 411.077694631718
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2709.57058128289
#>     AgeGV: 100 100.15 104.87 108.03 108.03 108.03 109.26 113.06 119 ...
#>     AgeGVp: 367.30328845583
#>     Autocut: NULL
#>     BM: 750 749.85825 745.39980939 742.456180652507 742.45618065 ...
#>     BMDR: 30 29.99325 29.78089779 29.6397363344754 29.639736334475 ...
#>     BMDRp: 0.646093312885173
#>     BMDV: 300 299.865 295.6189116 292.816444318032 292.81644431803 ...
#>     BMDVp: 270.430701507395
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 140
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 546.364532379602
#>     BM_after_cut: 1070
#>     ENV: 0 0 0 0 0 0 0 0 0.292 1 0 0 0.472 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0.292 1 0 0 0.472 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     LAIGV: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     OMD: 0.677265671641791 0.677283376707451 0.67784333368363 0.6 ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.859402985074627 0.859357089552239 0.857912910447761 0. ...
#>     OMDGR: 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0 ...
#>     OMDGV: 0.859402985074627 0.859357089552239 0.857912910447761 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 0 0.15 4.87 8.03 8.03 8.03 9.26 13.06 19.52 30.44 33.27  ...
#>     WR: 159.80079935245 160 160 160 159.518930436167 159.3773082 ...
#>     WRp: 155.407170432611
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 8724.35909045946
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 118 146 174 202 230 258 286 314
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 -0.14175 -4.45844061 -2.9436287374926 0 0 -1.135177993 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 98
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: sorens1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2015
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 56
#>     minBMGV: 184
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[2]][[4]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.288574824085632 0.248494987407072 0.272542889414208 0. ...
#>     AgeDR: 502.97 505.51 507.62 510.78 513.77 515.61 518.18 521.58  ...
#>     AgeDRp: 3276.08934295307
#>     AgeDV: 502.97 505.51 507.62 510.78 513.77 515.61 518.18 521.58  ...
#>     AgeDVp: 427.345709835846
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2492.27519480207
#>     AgeGV: 102.97 105.51 107.62 110.78 113.77 115.61 118.18 121.58  ...
#>     AgeGVp: 405.976163870246
#>     Autocut: NULL
#>     BM: 747.19335 744.8139274665 742.852022155526 739.9319657652 ...
#>     BMDR: 29.86635 29.7525592065 29.6583923566114 29.5178115768411 ...
#>     BMDRp: 0.34327925782018
#>     BMDV: 297.327 295.06136826 293.193629798914 290.41415418842 28 ...
#>     BMDVp: 262.620458905463
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 140
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 563.502994009661
#>     BM_after_cut: 1070
#>     ENV: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.11 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.11 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     LAIGV: 0.811104 0.811104 0.811104 0.811104 0.811104 0.811104 0. ...
#>     OMD: 0.677617482231559 0.677914037844372 0.678157294040934 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.858494253731343 0.857717089552239 0.857071492537313 0. ...
#>     OMDGR: 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0.89 0 ...
#>     OMDGV: 0.858494253731343 0.857717089552239 0.857071492537313 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.97 5.51 7.62 10.78 13.77 15.61 18.18 21.58 26.08 29.66 ...
#>     WR: 160 160 160 160 160 160 160 160 160 160 160 160 160 159. ...
#>     WRp: 140.565995243277
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 8624.43068429059
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 117 145 173 201 229 257 285
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: -2.80665 -2.3794225335 -1.96190531097437 -2.920056390264 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 88
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: sorens1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2016
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 56
#>     minBMGV: 184
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#>