Parameter Scan

Run ModVege for a different sets of parameters.

Usage

run_parameter_scan(environment, param_values, force = FALSE, outfilename = "")

Arguments

environment

Either a ModvegeEnvironment instance with all the site, management and weather inputs expected by ModvegeSite$run() or a string representing the name of a config file to read in order to generate the ModvegeEnvironment with read_config(). Note that, in the latter case, only the first found configuration is used if there are more than one valid uncommented lines in the config file.

param_values

A named list where each key stands for a ModVege parameter, i.e. a member of ModvegeParameters$parameter_names. Each list entry then has to be a vector containing the allowed values for the respective parameter. All possible allowed combinations of these parameter values are then generated and fed into a ModVege run.

force

Boolean. By default (force = FALSE), the function first counts the number of parameter combinations that need to be run and asks the user, if it should proceed. This can be suppressed by letting force = TRUE.

outfilename

String. If nonempty, the results are stored as an rds file with filename outfilename using the saveRDS() function.

Value

results A list containing an entry for each supplied parameter set in param_values. Each entry is itself a list containing the following keys:

params

The parameter set that was used to run ModVege for this entry.

data

A list containing for each simulated year a ModvegeSite object which was run for the respective year and therefore carries the respective results.

Note

Special care has to be taken in the creation of the param_values argument. It's possible to choose values that do not allow for any valid combination. Confer create_combinations().

See also

Examples

env = create_example_environment()
# We're creating a trivial list of parameters to explore here in order to 
# prevent the example from requiring a long time to execute. See 
# [create_combinations()] for more realistic uses of param_values.
param_values = list(w_FGA = c(0, 1), w_FGB = c(0, 1))
run_parameter_scan(env, param_values, force = TRUE)
#> [INFO]Parameter combination 1 out of 2.
#> [INFO]Starting run 1 out of 1.
#> [INFO][Run 1/1]Simulating year 2013 (1/4)
#> [INFO][Run 1/1]Simulating year 2014 (2/4)
#> [INFO][Run 1/1]Simulating year 2015 (3/4)
#> [INFO][Run 1/1]Simulating year 2016 (4/4)
#> [INFO]All runs completed.
#> [INFO]Parameter combination 2 out of 2.
#> [INFO]Starting run 1 out of 1.
#> [INFO][Run 1/1]Simulating year 2013 (1/4)
#> [INFO][Run 1/1]Simulating year 2014 (2/4)
#> [INFO][Run 1/1]Simulating year 2015 (3/4)
#> [INFO][Run 1/1]Simulating year 2016 (4/4)
#> [INFO]All runs completed.
#> [INFO]Completed parameter scan.
#> [INFO]Time used: 1.42987704277039
#> [[1]]
#> [[1]]$params
#> w_FGA w_FGB 
#>     0     1 
#> 
#> [[1]]$data
#> [[1]]$data[[1]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.457223659360304 0.492747302067495 0.480105581585783 0. ...
#>     AgeDR: 502.84 504.12 504.79 508.13 513.74 517.69 519.45 519.45  ...
#>     AgeDRp: 2656.25685098306
#>     AgeDV: 502.84 504.12 504.79 508.13 513.74 517.69 519.45 519.45  ...
#>     AgeDVp: 360.163681721118
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2535.539652403
#>     AgeGV: 102.84 104.12 104.79 108.13 113.74 117.69 119.45 119.45  ...
#>     AgeGVp: 418.243498780562
#>     BM: 748.2108 747.408871296 746.990160669531 744.905587078854 ...
#>     BMDR: 29.9148 29.876509056 29.8564917949325 29.7567711123374 2 ...
#>     BMDRp: 1.62661938055988
#>     BMDV: 298.296 297.53236224 297.133668874598 295.148815966516 2 ...
#>     BMDVp: 206.152935636412
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 459.313491565678
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     LAIGV: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     OMD: 0.678955476183984 0.678933962545162 0.678922359910314 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.861435 0.860955 0.86070375 0.85945125 0.8573475 0.8558 ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.861435 0.860955 0.86070375 0.85945125 0.8573475 0.8558 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.84 4.12 4.79 8.13 13.74 17.69 19.45 19.45 19.45 20.56  ...
#>     STp: 3324.38
#>     WR: 130 129.509787758099 129.068276627822 128.690056233788 1 ...
#>     WRp: 129.627111149557
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 6190.27925281306
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 106 134 162 190 218 246 274 302
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 95
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2013
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[1]]$data[[2]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.572832822953001 0.489068593905107 0.497218890323204 0. ...
#>     AgeDR: 502.07 505.44 508.61 512.35 515.58 517.36 524.65 532.15  ...
#>     AgeDRp: 2869.54048650331
#>     AgeDV: 502.07 505.44 508.61 512.35 515.58 517.36 524.65 532.15  ...
#>     AgeDVp: 423.673205672842
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2857.13439696621
#>     AgeGV: 102.07 105.44 108.61 112.35 115.58 117.36 124.65 132.15  ...
#>     AgeGVp: 485.543080985033
#>     BM: 748.6959 746.581380357 744.605437724945 742.288625723843 ...
#>     BMDR: 29.9379 29.837009277 29.7424259575919 29.6311892845105 2 ...
#>     BMDRp: 1.77999238727987
#>     BMDV: 298.758 296.74437108 294.863011767353 292.657436439333 2 ...
#>     BMDVp: 178.847655871668
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 625.73707954185
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0.458 0.5 0.454 0.184 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0.458 0.5 0.454 0.184 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     LAIGV: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     OMD: 0.678967568808644 0.678911882927523 0.678855169555004 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.86172375 0.86046 0.85927125 0.85786875 0.8566575 0.855 ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.86172375 0.86046 0.85927125 0.85786875 0.8566575 0.855 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.07 5.44 8.61 12.35 15.58 17.36 24.65 32.15 39.42 45.34 ...
#>     STp: 3719.5
#>     WR: 129.946718652693 130 130 130 130 129.365338865 128.74452 ...
#>     WRp: 127.0196850683
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 8625.14290032374
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 91 119 147 175 203 231 259 287
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 64
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2014
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[1]]$data[[3]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.00984744012611478
#>     ABSDV: 1.68970646494988
#>     AET: 0.19920064755 0.199001770202729 0.592140498337971 0.8132 ...
#>     AgeDR: 500 500 504.91 509.88 509.88 509.88 510.27 515.11 522.86 ...
#>     AgeDRp: 2965.34584618241
#>     AgeDV: 500 500 504.91 509.88 509.88 509.88 510.27 515.11 522.86 ...
#>     AgeDVp: 367.362080229803
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2914.28329749312
#>     AgeGV: 100 100 104.91 109.88 109.88 109.88 110.27 115.11 122.86 ...
#>     AgeGVp: 396.644510839453
#>     BM: 750 750 746.9067 743.805615321 743.805615321 743.8056153 ...
#>     BMDR: 30 30 29.8527 29.704332081 29.704332081 29.704332081 29. ...
#>     BMDRp: 1.55200001987625
#>     BMDV: 300 300 297.054 294.10128324 294.10128324 294.10128324 2 ...
#>     BMDVp: 199.728896566179
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 483.654227309786
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0 0 0.55 1 0 0 0.084 0.224 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0.55 1 0 0 0.084 0.224 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     LAIGV: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     OMD: 0.679 0.679 0.678922889565725 0.678834899723759 0.678834 ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.8625 0.8625 0.86065875 0.858795 0.858795 0.858795 0.85 ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.8625 0.8625 0.86065875 0.858795 0.858795 0.858795 0.85 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 0 0 4.91 9.88 9.88 9.88 10.27 15.11 22.86 35.36 40.23 42 ...
#>     STp: 3711.49
#>     WR: 129.800799840891 129.603644952911 130 129.709778732205 1 ...
#>     WRp: 123.422768956574
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 6590.8806478112
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 90 118 146 174 202 230 258 286 314
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 75
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2015
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[1]]$data[[4]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.272542889414208 0.224447085399936 0.288574824085632 0. ...
#>     AgeDR: 503.99 507.78 511.31 516.04 520.55 523.86 527.42 531.52  ...
#>     AgeDRp: 2869.83665126759
#>     AgeDV: 503.99 507.78 511.31 516.04 520.55 523.86 527.42 531.52  ...
#>     AgeDVp: 395.58725136375
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2715.92320594534
#>     AgeGV: 103.99 107.78 111.31 116.04 120.55 123.86 127.42 131.52  ...
#>     AgeGVp: 439.724338709058
#>     BM: 747.4863 745.117200183 742.926950449138 740.012362644197 ...
#>     BMDR: 29.8803 29.767053663 29.6619759635696 29.5216748172619 2 ...
#>     BMDRp: 1.06812089351777
#>     BMDV: 297.606 295.35014652 293.264974485569 290.490687826935 2 ...
#>     BMDVp: 192.642324000804
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 518.921428537923
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0 0 0 0 0.148 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0 0 0.148 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 4. ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     LAIGV: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     OMD: 0.678937386544743 0.678871776513771 0.678805235964392 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.86100375 0.8595825 0.85825875 0.856485 0.85479375 0.85 ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.86100375 0.8595825 0.85825875 0.856485 0.85479375 0.85 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 40.935 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 3.99 7.78 11.31 16.04 20.55 23.86 27.42 31.52 36.14 40.8 ...
#>     STp: 3500.48
#>     WR: 130 130 130 130 130 129.790711380507 130 130 130 130 130 ...
#>     WRp: 111.817550947756
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7428.39973800315
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 89 117 145 173 201 229 257 285
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 26
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2016
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> 
#> 
#> [[2]]
#> [[2]]$params
#> w_FGA w_FGB 
#>     1     0 
#> 
#> [[2]]$data
#> [[2]]$data[[1]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.457223659360304 0.492747302067495 0.480257065079617 0. ...
#>     AgeDR: 502.84 504.12 504.79 508.13 513.74 517.69 519.45 519.45  ...
#>     AgeDRp: 2610.42758488522
#>     AgeDV: 502.84 504.12 504.79 508.13 513.74 517.69 519.45 519.45  ...
#>     AgeDVp: 291.239024741609
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2576.82247323826
#>     AgeGV: 102.84 104.12 104.79 108.13 113.74 117.69 119.45 119.45  ...
#>     AgeGVp: 334.900531395507
#>     BM: 747.3588 746.178324096 745.562740186702 742.500042895222 ...
#>     BMDR: 29.9148 29.876509056 29.8564917949325 29.7567711123374 2 ...
#>     BMDRp: 2.58908004802842
#>     BMDV: 297.444 296.30181504 295.70624839177 292.743271782884 28 ...
#>     BMDVp: 224.509713889781
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 475.769909548837
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     LAIGV: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     OMD: 0.683551033318936 0.68370694231633 0.683788018117167 0.6 ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.869148 0.868764 0.868563 0.867561 0.865878 0.864693 0. ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.869148 0.868764 0.868563 0.867561 0.865878 0.864693 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.84 4.12 4.79 8.13 13.74 17.69 19.45 19.45 19.45 20.56  ...
#>     STp: 3324.38
#>     WR: 130 129.509787758099 129.068125144328 128.689534980198 1 ...
#>     WRp: 129.627111149557
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7344.92531937025
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 106 134 162 190 218 246 274 302
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 95
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2013
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[2]]$data[[2]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.572832822953001 0.489085329341227 0.497218890323204 0. ...
#>     AgeDR: 502.07 505.44 508.61 512.35 515.58 517.36 524.65 532.15  ...
#>     AgeDRp: 3565.92033154979
#>     AgeDV: 502.07 505.44 508.61 512.35 515.58 517.36 524.65 532.15  ...
#>     AgeDVp: 345.93062865456
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 3090.12454873003
#>     AgeGV: 102.07 105.44 108.61 112.35 115.58 117.36 124.65 132.15  ...
#>     AgeGVp: 417.770431170538
#>     BM: 748.0749 744.959844207 742.058642727408 738.667618102169 ...
#>     BMDR: 29.9379 29.837009277 29.7424259575919 29.6311892845105 2 ...
#>     BMDRp: 0.24682211890885
#>     BMDV: 298.137 295.12283493 292.316216769816 289.036428817658 2 ...
#>     BMDVp: 204.423245844888
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 734.455424879952
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0.458 0.5 0.454 0.184 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0.458 0.5 0.454 0.184 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     LAIGV: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     OMD: 0.683455613869681 0.683868591563627 0.684251053344278 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.869379 0.868368 0.867417 0.866295 0.865326 0.864792 0. ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.869379 0.868368 0.867417 0.866295 0.865326 0.864792 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.07 5.44 8.61 12.35 15.58 17.36 24.65 32.15 39.42 45.34 ...
#>     STp: 3719.5
#>     WR: 129.946718652693 130 130 130 130 129.365338865 128.74421 ...
#>     WRp: 127.007540797651
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 11994.4483863024
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 91 119 147 175 203 231 259 287
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 64
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2014
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[2]]$data[[3]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.00977550265766102
#>     ABSDV: 1.885032651351
#>     AET: 0.19920064755 0.199027247684387 0.592291422500415 0.8132 ...
#>     AgeDR: 500 500 504.91 509.88 509.88 509.88 510.27 515.11 522.86 ...
#>     AgeDRp: 2949.90378303938
#>     AgeDV: 500 500 504.91 509.88 509.88 509.88 510.27 515.11 522.86 ...
#>     AgeDVp: 300.918299167648
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2955.57629215679
#>     AgeGV: 100 100 104.91 109.88 109.88 109.88 110.27 115.11 122.86 ...
#>     AgeGVp: 319.174882513544
#>     BM: 750 750 745.4337 740.878219371 740.878219371 740.8782193 ...
#>     BMDR: 30 30 29.8527 29.704332081 29.704332081 29.704332081 29. ...
#>     BMDRp: 1.54066235739338
#>     BMDV: 300 300 295.581 291.17388729 291.17388729 291.17388729 2 ...
#>     BMDVp: 222.817098268439
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 518.292866942952
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0 0 0.55 1 0 0 0.084 0.224 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0.55 1 0 0 0.084 0.224 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     LAIGV: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     OMD: 0.6832 0.6832 0.683808459424359 0.684410863830488 0.6844 ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.87 0.87 0.868527 0.867036 0.867036 0.867036 0.866919 0 ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.87 0.87 0.868527 0.867036 0.867036 0.867036 0.866919 0 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 0 0 4.91 9.88 9.88 9.88 10.27 15.11 22.86 35.36 40.23 42 ...
#>     STp: 3711.49
#>     WR: 129.800799840891 129.603619475429 130 129.709778732205 1 ...
#>     WRp: 115.804802981945
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7654.69050450178
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 90 118 146 174 202 230 258 286 314
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 75
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2015
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[2]]$data[[4]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.272542889414208 0.224447085399936 0.288574824085632 0. ...
#>     AgeDR: 503.99 507.78 511.31 516.04 520.55 523.86 527.42 531.52  ...
#>     AgeDRp: 2797.3160776649
#>     AgeDV: 503.99 507.78 511.31 516.04 520.55 523.86 527.42 531.52  ...
#>     AgeDVp: 320.902844955838
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2770.13881329056
#>     AgeGV: 103.99 107.78 111.31 116.04 120.55 123.86 127.42 131.52  ...
#>     AgeGVp: 363.739785408334
#>     BM: 746.2893 742.805883333 739.597524427364 735.343037848355 ...
#>     BMDR: 29.8803 29.767053663 29.6619759635696 29.5216748172619 2 ...
#>     BMDRp: 1.94108326472162
#>     BMDV: 296.409 293.03882967 289.935548463795 285.821363031093 2 ...
#>     BMDVp: 217.757040002577
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 567.855133520762
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0 0 0 0 0.148 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0 0 0.148 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 5. ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     LAIGV: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     OMD: 0.683693883859785 0.684154536493993 0.684576016646551 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.868803 0.867666 0.866607 0.865188 0.863835 0.862842 0. ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.868803 0.867666 0.866607 0.865188 0.863835 0.862842 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 50.741 ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 3.99 7.78 11.31 16.04 20.55 23.86 27.42 31.52 36.14 40.8 ...
#>     STp: 3500.48
#>     WR: 130 130 130 130 130 129.790711380507 130 130 130 130 130 ...
#>     WRp: 111.634545336687
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 8860.71069775476
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 89 117 145 173 201 229 257 285
#>     cut_delays: 0
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_automatically: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     dry_days_after_cut: 2
#>     dry_days_before_cut: 1
#>     dry_precipitation_limit: 1
#>     dry_window: NULL
#>     end_of_cutting_season: NULL
#>     get_management: function () 
#>     get_target_biomass: function (DOY, intensity = "high") 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 26
#>     last_DOY_for_initial_cut: 150
#>     management: list
#>     max_cut_delay: 5
#>     max_cut_period: 55
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     stubble_heigt: 0.05
#>     target_biomass: NULL
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2016
#>   Private:
#>     REP_ON: 0.653846153846154
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (critical_temperature = 5, min_window_temperature = 6, 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> 
#>