Parameter Scan

Run ModVege for a different sets of parameters.

Usage

run_parameter_scan(environment, param_values, force = FALSE, outfilename = "")

Arguments

environment

Either a ModvegeEnvironment instance with all the site, management and weather inputs expected by ModvegeSite$run() or a string representing the name of a config file to read in order to generate the ModvegeEnvironment with read_config(). Note that, in the latter case, only the first found configuration is used if there are more than one valid uncommented lines in the config file.

param_values

A named list where each key stands for a ModVege parameter, i.e. a member of ModvegeParameters$parameter_names. Each list entry then has to be a vector containing the allowed values for the respective parameter. All possible allowed combinations of these parameter values are then generated and fed into a ModVege run.

force

Boolean. By default (force = FALSE), the function first counts the number of parameter combinations that need to be run and asks the user, if it should proceed. This can be suppressed by letting force = TRUE.

outfilename

String. If nonempty, the results are stored as an rds file with filename outfilename using the saveRDS() function.

Value

results A list containing an entry for each supplied parameter set in param_values. Each entry is itself a list containing the following keys:

params

The parameter set that was used to run ModVege for this entry.

data

A list containing for each simulated year a ModvegeSite object which was run for the respective year and therefore carries the respective results.

Note

Special care has to be taken in the creation of the param_values argument. It's possible to choose values that do not allow for any valid combination. Confer create_combinations().

See also

Examples

env = create_example_environment()
# We're creating a trivial list of parameters to explore here in order to 
# prevent the example from requiring a long time to execute. See 
# [create_combinations()] for more realistic uses of param_values.
param_values = list(w_FGA = c(0, 1), w_FGB = c(0, 1))
run_parameter_scan(env, param_values, force = TRUE)
#> [INFO]Parameter combination 1 out of 2.
#> [INFO]Starting run 1 out of 1.
#> [INFO] site name: `posieux1`.
#> [INFO]  run name: `-`.
#> [INFO][Run 1/1]Simulating year 2013 (1/4)
#> [INFO][Run 1/1]Simulating year 2014 (2/4)
#> [INFO][Run 1/1]Simulating year 2015 (3/4)
#> [INFO][Run 1/1]Simulating year 2016 (4/4)
#> [INFO]All runs completed.
#> [INFO]Parameter combination 2 out of 2.
#> [INFO]Starting run 1 out of 1.
#> [INFO] site name: `posieux1`.
#> [INFO]  run name: `-`.
#> [INFO][Run 1/1]Simulating year 2013 (1/4)
#> [INFO][Run 1/1]Simulating year 2014 (2/4)
#> [INFO][Run 1/1]Simulating year 2015 (3/4)
#> [INFO][Run 1/1]Simulating year 2016 (4/4)
#> [INFO]All runs completed.
#> [INFO]Completed parameter scan.
#> [INFO]Time used: 1.59371662139893
#> [[1]]
#> [[1]]$params
#> w_FGA w_FGB 
#>     0     1 
#> 
#> [[1]]$data
#> [[1]]$data[[1]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.457223659360304 0.492747302067495 0.480105581585783 0. ...
#>     AgeDR: 502.84 504.12 504.79 508.13 513.74 517.69 519.45 519.45  ...
#>     AgeDRp: 2644.26336625936
#>     AgeDV: 502.84 504.12 504.79 508.13 513.74 517.69 519.45 519.45  ...
#>     AgeDVp: 360.161532674197
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2535.539652403
#>     AgeGV: 102.84 104.12 104.79 108.13 113.74 117.69 119.45 119.45  ...
#>     AgeGVp: 418.286754052203
#>     Autocut: NULL
#>     BM: 748.1682 747.347207616 746.918550330996 744.784475166817 ...
#>     BMDR: 29.8722 29.814845376 29.7848814563971 29.6356592003006 2 ...
#>     BMDRp: 1.44984500161243
#>     BMDV: 298.296 297.53236224 297.133668874598 295.148815966516 2 ...
#>     BMDVp: 206.162374934968
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 459.312114185258
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     LAIGV: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     OMD: 0.678971359648806 0.678956977409514 0.678949101413265 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.861435 0.860955 0.86070375 0.85945125 0.8573475 0.8558 ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.861435 0.860955 0.86070375 0.85945125 0.8573475 0.8558 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.84 4.12 4.79 8.13 13.74 17.69 19.45 19.45 19.45 20.56  ...
#>     WR: 130 129.509787758099 129.068276627822 128.690056233788 1 ...
#>     WRp: 129.627111149557
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 6187.80616770428
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 106 134 162 190 218 246 274 302
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: -1.8318 -0.820992384 -0.42865728500448 -2.13407516417887 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 99
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2013
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 60
#>     minBMGV: 170
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[1]]$data[[2]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.572832822953001 0.489068593905107 0.497218890323204 0. ...
#>     AgeDR: 502.07 505.44 508.61 512.35 515.58 517.36 524.65 532.15  ...
#>     AgeDRp: 2863.25154199827
#>     AgeDV: 502.07 505.44 508.61 512.35 515.58 517.36 524.65 532.15  ...
#>     AgeDVp: 423.673204707313
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2857.13439696621
#>     AgeGV: 102.07 105.44 108.61 112.35 115.58 117.36 124.65 132.15  ...
#>     AgeGVp: 485.543081243553
#>     Autocut: NULL
#>     BM: 748.66485 746.50004195325 744.477194425601 742.105483532 ...
#>     BMDR: 29.90685 29.75567087325 29.6141826582477 29.448047093534 ...
#>     BMDRp: 1.61809910997691
#>     BMDV: 298.758 296.74437108 294.863011767353 292.657436439333 2 ...
#>     BMDVp: 178.847655912016
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 625.737079311742
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0.458 0.5 0.454 0.184 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0.458 0.5 0.454 0.184 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     LAIGV: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     OMD: 0.678979138662647 0.678942273076309 0.678903205018348 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.86172375 0.86046 0.85927125 0.85786875 0.8566575 0.855 ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.86172375 0.86046 0.85927125 0.85786875 0.8566575 0.855 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.07 5.44 8.61 12.35 15.58 17.36 24.65 32.15 39.42 45.34 ...
#>     WR: 129.946718652693 130 130 130 130 129.365338865 128.74452 ...
#>     WRp: 127.019685068308
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 8625.57488717009
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 91 119 147 175 203 231 259 287
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: -1.33515 -2.16480804675 -2.0228475276495 -2.371710892732 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 67
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2014
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 60
#>     minBMGV: 170
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[1]]$data[[3]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.00897994410067624
#>     ABSDV: 1.68970646494988
#>     AET: 0.19920064755 0.199001770202729 0.592140498337971 0.8132 ...
#>     AgeDR: 500 500 504.91 509.88 509.88 509.88 510.27 515.11 522.86 ...
#>     AgeDRp: 2956.80782663374
#>     AgeDV: 500 500 504.91 509.88 509.88 509.88 510.27 515.11 522.86 ...
#>     AgeDVp: 367.362080229803
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2914.28329749312
#>     AgeGV: 100 100 104.91 109.88 109.88 109.88 110.27 115.11 122.86 ...
#>     AgeGVp: 396.644510839453
#>     Autocut: NULL
#>     BM: 750 750 746.83305 743.65833042225 743.65833042225 743.65 ...
#>     BMDR: 30 30 29.77905 29.55704718225 29.55704718225 29.55704718 ...
#>     BMDRp: 1.41527881807348
#>     BMDV: 300 300 297.054 294.10128324 294.10128324 294.10128324 2 ...
#>     BMDVp: 199.728896566179
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 483.654227309786
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0 0 0.55 1 0 0 0.084 0.224 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0.55 1 0 0 0.084 0.224 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     LAIGV: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     OMD: 0.679 0.679 0.678950395941904 0.678890124237886 0.678890 ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.8625 0.8625 0.86065875 0.858795 0.858795 0.858795 0.85 ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.8625 0.8625 0.86065875 0.858795 0.858795 0.858795 0.85 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 0 0 4.91 9.88 9.88 9.88 10.27 15.11 22.86 35.36 40.23 42 ...
#>     WR: 129.800799840891 129.603644952911 130 129.709778732205 1 ...
#>     WRp: 123.422768956574
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 6589.80768432111
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 90 118 146 174 202 230 258 286 314
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 -3.16695 -3.17471957775 0 0 -0.246689873528816 -3.05 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 75
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2015
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 60
#>     minBMGV: 170
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[1]]$data[[4]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.272542889414208 0.224447085399936 0.288574824085632 0. ...
#>     AgeDR: 503.99 507.78 511.31 516.04 520.55 523.86 527.42 531.52  ...
#>     AgeDRp: 2855.49320188429
#>     AgeDV: 503.99 507.78 511.31 516.04 520.55 523.86 527.42 531.52  ...
#>     AgeDVp: 395.587096271218
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2715.92320594534
#>     AgeGV: 103.99 107.78 111.31 116.04 120.55 123.86 127.42 131.52  ...
#>     AgeGVp: 439.724446778537
#>     Autocut: NULL
#>     BM: 747.42645 745.00106726175 742.758893601991 739.775347587 ...
#>     BMDR: 29.82045 29.65092074175 29.4939191164224 29.284659760291 ...
#>     BMDRp: 0.955716854950983
#>     BMDV: 297.606 295.35014652 293.264974485569 290.490687826935 2 ...
#>     BMDVp: 192.642331994861
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 518.921400037853
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0 0 0 0 0.148 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0 0 0.148 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     LAIGV: 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0.714 0. ...
#>     OMD: 0.678959722391414 0.678915247852919 0.678868318519617 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.86100375 0.8595825 0.85825875 0.856485 0.85479375 0.85 ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.86100375 0.8595825 0.85825875 0.856485 0.85479375 0.85 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 3.99 7.78 11.31 16.04 20.55 23.86 27.42 31.52 36.14 40.8 ...
#>     WR: 130 130 130 130 130 129.790711380507 130 130 130 130 130 ...
#>     WRp: 111.817550969167
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7384.5747243203
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 89 117 145 173 201 229 257 285
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: -2.57355 -2.42538273825 -2.24217365975877 -2.98354601476 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 85
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2016
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 60
#>     minBMGV: 170
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> 
#> 
#> [[2]]
#> [[2]]$params
#> w_FGA w_FGB 
#>     1     0 
#> 
#> [[2]]$data
#> [[2]]$data[[1]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.457223659360304 0.492747302067495 0.480257065079617 0. ...
#>     AgeDR: 502.84 504.12 504.79 508.13 513.74 517.69 519.45 519.45  ...
#>     AgeDRp: 2604.47104624449
#>     AgeDV: 502.84 504.12 504.79 508.13 513.74 517.69 519.45 519.45  ...
#>     AgeDVp: 291.237590236816
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2576.82254254577
#>     AgeGV: 102.84 104.12 104.79 108.13 113.74 117.69 119.45 119.45  ...
#>     AgeGVp: 334.913135791325
#>     Autocut: NULL
#>     BM: 747.3162 746.116660416 745.491129848167 742.378930983185 ...
#>     BMDR: 29.8722 29.814845376 29.7848814563971 29.6356592003006 2 ...
#>     BMDRp: 2.3678036809523
#>     BMDV: 297.444 296.30181504 295.70624839177 292.743271782884 28 ...
#>     BMDVp: 224.510231190874
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 475.768631032088
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0.122 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     LAIGV: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     OMD: 0.683567196857234 0.683730389606859 0.683815278208181 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.869148 0.868764 0.868563 0.867561 0.865878 0.864693 0. ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.869148 0.868764 0.868563 0.867561 0.865878 0.864693 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.84 4.12 4.79 8.13 13.74 17.69 19.45 19.45 19.45 20.56  ...
#>     WR: 130 129.509787758099 129.068125144328 128.689534980198 1 ...
#>     WRp: 129.627111149557
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7336.81826611982
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 106 134 162 190 218 246 274 302
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: -2.6838 -1.199539584 -0.62553056783328 -3.11219886498208 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 99
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2013
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 60
#>     minBMGV: 170
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[2]]$data[[2]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.572832822953001 0.489085329341227 0.497218890323204 0. ...
#>     AgeDR: 502.07 505.44 508.61 512.35 515.58 517.36 524.65 532.15  ...
#>     AgeDRp: 3550.6203981573
#>     AgeDV: 502.07 505.44 508.61 512.35 515.58 517.36 524.65 532.15  ...
#>     AgeDVp: 345.930628643749
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 3090.12465354209
#>     AgeGV: 102.07 105.44 108.61 112.35 115.58 117.36 124.65 132.15  ...
#>     AgeGVp: 417.770431170618
#>     Autocut: NULL
#>     BM: 748.04385 744.87850580325 741.930399428063 738.484475911 ...
#>     BMDR: 29.90685 29.75567087325 29.6141826582477 29.448047093534 ...
#>     BMDRp: 0.229773434302287
#>     BMDV: 298.137 295.12283493 292.316216769816 289.036428817658 2 ...
#>     BMDVp: 204.42324584491
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 734.455424879798
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0.458 0.5 0.454 0.184 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0.458 0.5 0.454 0.184 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     LAIGV: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     OMD: 0.683467379619524 0.683899589126226 0.684300186380141 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.869379 0.868368 0.867417 0.866295 0.865326 0.864792 0. ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.869379 0.868368 0.867417 0.866295 0.865326 0.864792 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 2.07 5.44 8.61 12.35 15.58 17.36 24.65 32.15 39.42 45.34 ...
#>     WR: 129.946718652693 130 130 130 130 129.365338865 128.74421 ...
#>     WRp: 127.007540797651
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 11994.9123270293
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 91 119 147 175 203 231 259 287
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: -1.95615 -3.16534419675 -2.9481063751866 -3.445923516870 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 67
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2014
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 60
#>     minBMGV: 170
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[2]]$data[[3]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0.00881995527651962
#>     ABSDV: 1.885032651351
#>     AET: 0.19920064755 0.199027247684387 0.592291422500415 0.8132 ...
#>     AgeDR: 500 500 504.91 509.88 509.88 509.88 510.27 515.11 522.86 ...
#>     AgeDRp: 2946.0617402844
#>     AgeDV: 500 500 504.91 509.88 509.88 509.88 510.27 515.11 522.86 ...
#>     AgeDVp: 300.918299167648
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2955.57629215679
#>     AgeGV: 100 100 104.91 109.88 109.88 109.88 110.27 115.11 122.86 ...
#>     AgeGVp: 319.174882513544
#>     Autocut: NULL
#>     BM: 750 750 745.36005 740.73093447225 740.73093447225 740.73 ...
#>     BMDR: 30 30 29.77905 29.55704718225 29.55704718225 29.55704718 ...
#>     BMDRp: 1.39006387336795
#>     BMDV: 300 300 295.581 291.17388729 291.17388729 291.17388729 2 ...
#>     BMDVp: 222.817098268439
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 518.292866942952
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0 0 0.55 1 0 0 0.084 0.224 0 0 0 0 0 0 0 0 0 ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0.55 1 0 0 0.084 0.224 0 0 0 0 0 0 0 0 0 ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     LAIGV: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     OMD: 0.6832 0.6832 0.683836502908896 0.684467415303274 0.6844 ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.87 0.87 0.868527 0.867036 0.867036 0.867036 0.866919 0 ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.87 0.87 0.868527 0.867036 0.867036 0.867036 0.866919 0 ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 0 0 4.91 9.88 9.88 9.88 10.27 15.11 22.86 35.36 40.23 42 ...
#>     WR: 129.800799840891 129.603619475429 130 129.709778732205 1 ...
#>     WRp: 115.804802981945
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 7653.37466453516
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 90 118 146 174 202 230 258 286 314
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: 0 0 -4.63995 -4.62911552775 0 0 -0.357964320730916 -4.43 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 75
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2015
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 60
#>     minBMGV: 170
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> [[2]]$data[[4]]
#> <ModvegeSite>
#>   Public:
#>     ABSDR: 0
#>     ABSDV: 0
#>     AET: 0.272542889414208 0.224447085399936 0.288574824085632 0. ...
#>     AgeDR: 503.99 507.78 511.31 516.04 520.55 523.86 527.42 531.52  ...
#>     AgeDRp: 2793.25231684086
#>     AgeDV: 503.99 507.78 511.31 516.04 520.55 523.86 527.42 531.52  ...
#>     AgeDVp: 320.902844701959
#>     AgeGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     AgeGRp: 2770.18232764795
#>     AgeGV: 103.99 107.78 111.31 116.04 120.55 123.86 127.42 131.52  ...
#>     AgeGVp: 363.739785300284
#>     Autocut: NULL
#>     BM: 746.22945 742.68975041175 739.429467580217 735.106022791 ...
#>     BMDR: 29.82045 29.65092074175 29.4939191164224 29.284659760291 ...
#>     BMDRp: 1.76859683415131
#>     BMDV: 296.409 293.03882967 289.935548463795 285.821363031093 2 ...
#>     BMDVp: 217.757040012907
#>     BMG: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGR: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     BMGRp: 150
#>     BMGV: 420 420 420 420 420 420 420 420 420 420 420 420 420 420  ...
#>     BMGVp: 567.855133756141
#>     BM_after_cut: 900
#>     ENV: 0 0 0 0 0 0 0 0 0 0 0.148 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfPAR: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     ENVfT: 0 0 0 0 0 0 0 0 0 0 0.148 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     ENVfW: 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1  ...
#>     GRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     GROGR: 0
#>     GROGV: 0
#>     LAI: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     LAIGV: 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248 0.94248  ...
#>     OMD: 0.683716637020959 0.684198969174519 0.684640694821047 0. ...
#>     OMDDR: NULL
#>     OMDDV: NULL
#>     OMDG: 0.868803 0.867666 0.866607 0.865188 0.863835 0.862842 0. ...
#>     OMDGR: 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9  ...
#>     OMDGV: 0.868803 0.867666 0.866607 0.865188 0.863835 0.862842 0. ...
#>     PGRO: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     REP: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     SENGR: 0
#>     SENGV: 0
#>     ST: 3.99 7.78 11.31 16.04 20.55 23.86 27.42 31.52 36.14 40.8 ...
#>     WR: 130 130 130 130 130 129.790711380507 130 130 130 130 130 ...
#>     WRp: 111.634545336509
#>     cBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     cBMp: 8766.41508046534
#>     clone: function (deep = FALSE) 
#>     cut_DOYs: 89 117 145 173 201 229 257 285
#>     cut_during_growth_preriod: TRUE
#>     cut_height: 0.05
#>     dBM: -3.77055 -3.53969958825 -3.26028283153287 -4.32344478883 ...
#>     days_per_year: 365
#>     determine_cut: function (DOY) 
#>     determine_cut_from_input: function (DOY) 
#>     get_management: function () 
#>     get_weather: function () 
#>     hvBM: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0  ...
#>     initialize: function (parameters, site_name = "-", run_name = "-") 
#>     j_start_of_growing_season: 85
#>     management: list
#>     n_state_variables: 31
#>     parameters: ModvegeParameters, R6
#>     plot: function (...) 
#>     plot_bm: function (smooth_interval = 28, ...) 
#>     plot_growth: function (...) 
#>     plot_limitations: function (...) 
#>     plot_var: function (var, ...) 
#>     plot_water: function (...) 
#>     run: function (year, weather, management) 
#>     run_name: -
#>     set_SGS_method: function (method) 
#>     set_parameters: function (params) 
#>     site_name: posieux1
#>     state_variable_names: AgeGV AgeGR AgeDV AgeDR BMGV BMGR BMDV BMDR OMDGV OMDGR  ...
#>     time_step: 1
#>     version: package_version, numeric_version
#>     weather: list
#>     write_output: function (filename, force = FALSE) 
#>     year: 2016
#>   Private:
#>     REP_ON: 0.653846153846154
#>     SGS_method: MTD
#>     SGS_options: MTD simple
#>     apply_cuts: function () 
#>     calculate_ageing: function () 
#>     calculate_digestibility: function () 
#>     calculate_growth: function () 
#>     calculate_temperature_sum: function () 
#>     carry_over_from_last_day: function () 
#>     check_if_simulation_has_run: function () 
#>     current_DOY: 365
#>     get_start_of_growing_season: function (first_possible_DOY = 30, consider_snow = FALSE, critical_snow = 1) 
#>     initialize_state_variables: function () 
#>     make_header: function () 
#>     minBMGR: 60
#>     minBMGV: 170
#>     update_biomass: function () 
#>     vars_to_exclude: OMDDV OMDDR
#>     ylabels: list
#> 
#> 
#>